+++ paper/paper.bib
@@ -88,14 +88,21 @@
 	journal = {Progress in Oceanography},
 	pages = {176-193},
 	title = {Model investigations of the North Atlantic spring bloom initiation},
-	url = {https://www.sciencedirect.com/science/article/pii/S0079661115001561},
 	volume = {138},
-	year = {2015},
-	bdsk-url-1 = {https://www.sciencedirect.com/science/article/pii/S0079661115001561},
-	bdsk-url-2 = {https://doi.org/10.1016/j.pocean.2015.07.004}
+	year = {2015}
 }
 
-@article{ekp, doi = {10.21105/joss.04869}, url = {https://doi.org/10.21105/joss.04869}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {80}, pages = {4869}, author = {Oliver R. a. Dunbar and Ignacio Lopez-Gomez and Alfredo Garbuno-Iñigo and Daniel Zhengyu Huang and Eviatar Bach and Jin-long Wu}, title = {EnsembleKalmanProcesses.jl: Derivative-free ensemble-based model calibration}, journal = {Journal of Open Source Software} } 
+@article{ekp,
+   doi = {10.21105/joss.04869},
+   url = {https://doi.org/10.21105/joss.04869},
+   year = {2022},
+   publisher = {The Open Journal},
+   volume = {7},
+   number = {80},
+   pages = {4869},
+   author = {Oliver R. a. Dunbar and Ignacio Lopez-Gomez and Alfredo Garbuno-Iñigo and Daniel Zhengyu Huang and Eviatar Bach and Jin-long Wu},
+   title = {EnsembleKalmanProcesses.jl: Derivative-free ensemble-based model calibration},
+   journal = {Journal of Open Source Software}} 
 
 
 @article{pisces,
@@ -110,8 +117,7 @@
 	publisher = {Copernicus GmbH},
 	title = {PISCES-v2: An ocean biogeochemical model for carbon and ecosystem studies},
 	volume = {8},
-	year = {2015},
-	bdsk-url-1 = {https://doi.org/10.5194/gmd-8-2465-2015}}
+	year = {2015}}
 
 
 @article{broch:2012,
@@ -126,8 +132,7 @@
 	pages = {759-776},
 	title = {Modelling seasonal growth and composition of the kelp Saccharina latissima},
 	volume = {24},
-	year = {2012},
-	bdsk-url-1 = {https://doi.org/10.1007/s10811-011-9695-y}}
+	year = {2012}}
 
 
 @article{julia,
@@ -135,6 +140,7 @@
 	journal = {SIAM Review},
 	number = {1},
 	pages = {65-98},
-	title = {Julia: A Fresh Approach to Numerical Computing},
+	title = {Julia: A fresh approach to numerical computing},
 	volume = {59},
-	year = {2017}}
+	year = {2017},
+   doi = {10.1137/141000671}}
