+++ paper/paper.md
@@ -14,6 +14,7 @@ authors:
   - name: Si Chen
     affiliation: "1, 2"
   - name: Navid C Constantinou
+    orcid: 0000-0002-8149-4094
     affiliation: 3, 4
   - name: Simone Silvestri
     affiliation: 5
