### 2023-07-04 77ded36
changed `xnode` use to `xnodes`
Not sure this is the most efficient way todo this, but since `xnode` is defiend differedntly for different grid
types this is by far the easiest fix. The alternative would be to define `xnode(i, grid, lx, ly, lz)` for RectilinearGrids.
Perhaps that solution would be better.
