### 2023-07-06 c530081
code cleanup
