### 2023-07-06 a3b078b
tidied kelp model
