+++ src/Boundaries/Sediments/coupled_timesteppers.jl
@@ -1,13 +1,15 @@
-using Oceananigans: NonhydrostaticModel, prognostic_fields
+using Oceananigans: NonhydrostaticModel, prognostic_fields, HydrostaticFreeSurfaceModel
 using OceanBioME: ContinuousFormBiogeochemistry
 using OceanBioME.Boundaries.Sediments: AbstractSediment
 using Oceananigans.TimeSteppers: ab2_step_field!, rk3_substep_field!, stage_Δt
 using Oceananigans.Utils: work_layout, launch!
 using Oceananigans.TurbulenceClosures: implicit_step!
+using Oceananigans.Models.HydrostaticFreeSurfaceModels: local_ab2_step!, ab2_step_free_surface!
+using Oceananigans.Architectures: AbstractArchitecture
 
 import Oceananigans.TimeSteppers: ab2_step!, rk3_substep!
 
-@inline function ab2_step!(model::NonhydrostaticModel{<:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:ContinuousFormBiogeochemistry{<:Any, <:FlatSediment}}, Δt, χ)
+@inline function ab2_step!(model::NonhydrostaticModel{<:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:ContinuousFormBiogeochemistry{<:Any, <:FlatSediment}}, Δt, χ) 
     workgroup, worksize = work_layout(model.grid, :xyz)
     arch = model.architecture
     step_field_kernel! = ab2_step_field!(device(arch), workgroup, worksize)
@@ -42,6 +44,25 @@ import Oceananigans.TimeSteppers: ab2_step!, rk3_substep!
     end
 
     return nothing
+end
+
+@inline function ab2_step!(model::HydrostaticFreeSurfaceModel{<:Any, <:Any, <:AbstractArchitecture, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:Any, <:ContinuousFormBiogeochemistry{<:Any, <:FlatSediment}}, Δt, χ)
+    # Step locally velocity and tracers
+    @apply_regionally local_ab2_step!(model, Δt, χ)
+
+    # blocking step for implicit free surface, non blocking for explicit
+    ab2_step_free_surface!(model.free_surface, model, Δt, χ)
+
+    sediment = model.biogeochemistry.sediment_model
+    arch = model.architecture
+
+    for (i, field) in enumerate(sediment_fields(sediment))
+        launch!(arch, model.grid, :xy, ab2_step_flat_field!, 
+                field, Δt, χ, 
+                sediment.tendencies.Gⁿ[i], 
+                sediment.tendencies.G⁻[i])
+
+    end
 end
 
 @kernel function ab2_step_flat_field!(u, Δt, χ, Gⁿ, G⁻)
