### 2023-07-12 bfb86c7
added simple regression case
