### 2023-07-12 ca31941
run the tests
