+++ test/runtests.jl
@@ -5,3 +5,4 @@ include("test_NPZD.jl")
 include("test_gasexchange.jl")
 include("test_slatissima.jl")
 include("test_sediments.jl")
+include("regression_tests/run_regresisons.jl")
