+++ test/test_sediments.jl
@@ -1,11 +1,13 @@
 using OceanBioME, Oceananigans, Test, JLD2
 
-using OceanBioME.Sediments: SimpleMultiG
+using OceanBioME.Sediments: SimpleMultiG, InstantRemineralisation
 using Oceananigans.Units
 
 using OceanBioME.Sediments: sediment_tracers, sediment_fields
 using Oceananigans: Field
 using Oceananigans.Fields: TracerFields
+
+using Oceananigans.Operators: volume, Azᶠᶜᶜ
 
 function intercept_tendencies!(model, intercepted_tendencies)
     for tracer in keys(model.tracers)
@@ -13,48 +15,53 @@ function intercept_tendencies!(model, intercepted_tendencies)
     end
 end
 
-function test_flat_sediment(architecture; timestepper = :QuasiAdamsBashforth2)
-    grid = RectilinearGrid(architecture; size=(3, 3, 3), extent=(10, 10, 200))
+function set_defaults!(sediment::SimpleMultiG)
+    set!(sediment.fields.N_fast, 0.0230)
+    set!(sediment.fields.N_slow, 0.0807)
 
-    sediment_model = SimpleMultiG(grid)
+    set!(sediment.fields.C_fast, 0.5893)
+    set!(sediment.fields.C_slow, 0.1677)
+end 
 
-    biogeochemistry = LOBSTER(; grid, 
-                                carbonates = true, oxygen = true, variable_redfield = true, 
-                                open_bottom = true, 
-                                surface_phytosynthetically_active_radiation = (x, y, t) -> 80,
-                                sediment_model)
+set_defaults!(::InstantRemineralisation) = nothing
 
-    model = NonhydrostaticModel(;grid, biogeochemistry, 
-                                 closure = nothing,
-                                 advection = nothing,
-                                 timestepper)
-
-    sed_tracers = sediment_tracers(model.biogeochemistry.sediment_model)
-    @test all(tracer ∈ sed_tracers for tracer in (:C_slow, :C_fast, :C_ref, :N_slow, :N_fast, :N_ref))
-
-    sed_fields = sediment_fields(model.biogeochemistry.sediment_model)
-    @test isa(sed_fields.C_slow, Field)
-    @test all(tracer ∈ keys(sed_fields) for tracer in (:C_slow, :C_fast, :C_ref, :N_slow, :N_fast, :N_ref))
-
-    set!(model.biogeochemistry.sediment_model.fields.N_fast, 0.0230)
-    set!(model.biogeochemistry.sediment_model.fields.N_slow, 0.0807)
-
-    set!(model.biogeochemistry.sediment_model.fields.C_fast, 0.5893)
-    set!(model.biogeochemistry.sediment_model.fields.C_slow, 0.1677)
-
+function set_defaults!(::LOBSTER, model)
     set!(model, P = 0.4686, Z = 0.5363, 
-            NO₃ = 2.3103, NH₄ = 0.0010, 
-            DIC = 2106.9, Alk = 2408.9, 
-            O₂ = 258.92, 
-            DOC = 5.3390, DON = 0.8115,
-            sPON = 0.2299, sPOC = 1.5080,
-            bPON = 0.0103, bPOC = 0.0781)
+                NO₃ = 2.3103, NH₄ = 0.0010, 
+                DIC = 2106.9, Alk = 2408.9, 
+                O₂ = 258.92, 
+                DOC = 5.3390, DON = 0.8115,
+                sPON = 0.2299, sPOC = 1.5080,
+                bPON = 0.0103, bPOC = 0.0781)
+end
 
-    simulation = Simulation(model, Δt = 1, stop_iteration = 1)
[…]
