### 2023-07-13 32ccb3c
added test that sediment and particles work together
