+++ test/test_slatissima.jl
@@ -41,7 +41,7 @@ end
 
     initial_tracer_C = sum(model.tracers.sPOC) + sum(model.tracers.bPOC) + sum(model.tracers.DOC) + sum(model.tracers.DIC) + 
                        sum(model.tracers.P * (1 + model.biogeochemistry.organic_carbon_calcate_ratio) .+ model.tracers.Z) * model.biogeochemistry.phytoplankton_redfield 
-                       sum(model.tracers.sPOC) + sum(model.tracers.bPOC) + sum(model.tracers.DOC)
+                    
     initial_kelp_C = sum(particles.A .* particles.structural_dry_weight_per_area .* (particles.C .+ particles.structural_carbon)) ./ (12 * 0.001)
 
     model.clock.time = 60days # get to a high growth phase
