+++ docs/make.jl
@@ -53,7 +53,8 @@ model_parameters = (LOBSTER(; grid = BoxModelGrid()),
                     NutrientPhytoplanktonZooplanktonDetritus(; grid = BoxModelGrid()),
                     SLatissima(),
                     TwoBandPhotosyntheticallyActiveRadiation(; grid = BoxModelGrid()),
-                    SimpleMultiG(BoxModelGrid(); depth = 1000),
+                    SimpleMultiG(; grid = BoxModelGrid(), depth = 1000),
+                    InstantRemineralisation(; BoxModelGrid())
                     OCMIP_default,
                     GasExchange(; gas = :CO₂).condition.parameters,
                     GasExchange(; gas = :O₂).condition.parameters)
