### 2023-07-14 f0be386
Apply suggestions from code review
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>
