### 2023-07-15 9612180
convert SimpleMultiG into doctest
