+++ docs/src/model_components/sediment.md
@@ -4,13 +4,18 @@ We currently have one single-layer sediment model implemented. The model, propos
 
 It is straightforward to set up:
 
-```@example
-using OceanBioME, OceanBioME.Sediments
-using Oceananigans
+```jldoctest simplemultig; filter = r".*@ OceanBioME.Boundaries.Sediments.*"
+using OceanBioME, Oceananigans, OceanBioME.Sediments
 
 grid = RectilinearGrid(size=(3, 3, 30), extent=(10, 10, 200))
 
 sediment_model = SimpleMultiG(; grid)
+
+# output
+┌ Warning: Sediment models are an experimental feature and have not yet been validated.
+└ @ OceanBioME.Boundaries.Sediments ~/OceanBioME.jl/src/Boundaries/Sediments/simple_multi_G.jl:102
+[ Info: This sediment model is currently only compatible with models providing NH₄, NO₃, O₂, and DIC.
+Single-layer multi-G sediment model (Float64)
 ```
 
 You may optionally specify the model parameters. This can then be passed in the setup of a BGC model:
