### 2023-07-15 c62659e
run doctests as part of tests
