+++ test/runtests.jl
@@ -1,3 +1,5 @@
+using OceanBioME, Documenter, Test
+
 include("test_utils.jl")
 include("test_light.jl")
 include("test_LOBSTER.jl")
@@ -5,3 +7,7 @@ include("test_NPZD.jl")
 include("test_gasexchange.jl")
 include("test_slatissima.jl")
 include("test_sediments.jl")
+
+@testset "Doctests" begin
+    doctest(OceanBioME)
+end
