+++ test/test_utils.jl
@@ -1,14 +1,14 @@
 using Test, OceanBioME, Oceananigans
 
 function test_column_diffusion_timescale(arch)
-    κ = 1e-3
+    const κ = 1e-3
     @inline κₜ(x, y, z, t) = κ
 
     grid = RectilinearGrid(arch, size=(1, 1, 5), x=(0, 10), y=(0, 3), z=[-1, -0.6, -0.5, -0.2, -0.19, 0])
-    min_Δz = 1e-2 # for the grid above
+    min_Δz = minimum_zspacing(grid)
 
     model = NonhydrostaticModel(; grid,
-                                closure = ScalarDiffusivity(ν = κₜ, κ = κₜ), 
+                                closure = ScalarDiffusivity(ν = κₜ, κ = κₜ),
                                 biogeochemistry = LOBSTER(; grid,
                                                             surface_phytosynthetically_active_radiation = (x, y, t) -> 100,
                                                             carbonates = true),
@@ -27,7 +27,7 @@ function test_negative_scaling(arch)
     simulation = Simulation(model, Δt = 1)
 
     negativity_protection! = ScaleNegativeTracers(; model, tracers = (:A, :B))
-    
+
     negativity_protection!(simulation)
 
     return (model.tracers.A[1, 1, 1] == 0.5) && (model.tracers.B[1, 1, 1] == 0.0)
