+++ .github/workflows/tests.yml
@@ -22,11 +22,12 @@ jobs:
       fail-fast: false
       matrix:
         version:
+          - '1.9'
           - '1.8'
         os:
           - ubuntu-latest
         arch:
-          - x64          
+          - x64
     steps:
       - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@v1
@@ -46,6 +47,6 @@ jobs:
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
-      - uses: codecov/codecov-action@v1
+      - uses: codecov/codecov-action@v2
         with:
           file: lcov.info
