### 2023-07-15 61ea1c7
drop const from local var
