### 2023-07-15 2391f08
better test messages
testing on CPU() -> testing on CPU
