### 2023-07-15 1c53892
add version for styfle/cancel-workflow-action
