+++ .github/workflows/documentation.yml
@@ -21,7 +21,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Cancel Previous Runs
-        uses: styfle/cancel-workflow-action
+        uses: styfle/cancel-workflow-action@v0.11
         with:
           access_token: ${{ github.token }}
       - uses: actions/checkout@v3
