+++ .github/workflows/tests.yml
@@ -30,7 +30,7 @@ jobs:
           - x64
     steps:
       - name: Cancel Previous Runs
-        uses: styfle/cancel-workflow-action
+        uses: styfle/cancel-workflow-action@v0.11
       - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@v1
         with:
