+++ .github/workflows/documentation.yml
@@ -20,10 +20,6 @@ jobs:
       contents: write
     runs-on: ubuntu-latest
     steps:
-      - name: Cancel Previous Runs
-        uses: styfle/cancel-workflow-action@v0.11
-        with:
-          access_token: ${{ github.token }}
       - uses: actions/checkout@v3
       - uses: julia-actions/setup-julia@latest
         with:
