### 2023-07-15 3835006
minor cleanup + add missing dep for box_npzd
