### 2023-07-15 4802628
M² is the shear squared
