### 2023-07-15 d1fc31a
add more explanation for background state
