+++ examples/eady.jl
@@ -1,54 +1,58 @@
 # # Biogeochemistry in submesoscale eddies in the Eady model
-# In this example we will setup a 3D model with a constant background buoyancy gradient with associated 
-# thermal wind (the [Eady model](https://en.wikipedia.org/wiki/Eady_model)) with the [LOBSTER](@ref LOBSTER) biogeochemical model. 
-# This demonstrates how to use biogeochemistry in a more complicated physical model.
-# The parameters correspond roughly to those in [taylor2016](@cite) and will generate a single submesoscale eddy
+#
+# In this example we setup a 3D model with a constant background buoyancy gradient with associated 
+# thermal wind (the [Eady model](https://en.wikipedia.org/wiki/Eady_model)) with the [LOBSTER](@ref LOBSTER)
+# biogeochemical model. This demonstrates how to use biogeochemistry in a more complicated physical model.
+# The parameters in this example correspond roughly to those used by [taylor2016](@cite) and lead to the
+# generation of a single submesoscale eddy.
 
 # ## Install dependencies
-# First we will check we have the dependencies installed
+# First we ensure we have the required dependencies installed
 # ```julia
 # using Pkg
 # pkg "add OceanBioME, Oceananigans, CairoMakie"
 # ```
 
 # ## Model setup
-# First load the required packages. We also set the random seed to ensure reproducibility of the results. (not required)
+# We load the required packages. Although not required, we also set the random seed to ensure
+# reproducibility of the results.
 using OceanBioME, Oceananigans, Printf
 using Oceananigans.Units
 using Random
 
-Random.seed!(42);
+Random.seed!(1234)
 
-# Construct a grid with uniform grid spacing
-grid = RectilinearGrid(size=(32, 32, 8), extent = (1kilometer, 1kilometer, 100meters))
+# Construct a grid with uniform grid spacing.
+grid = RectilinearGrid(size = (32, 32, 8), extent = (1kilometer, 1kilometer, 100meters))
 
-# Set the Coriolis parameter
+# Set the Coriolis parameter.
 coriolis = FPlane(f = 1e-4) # [s⁻¹]
 
-# Specify parameters that are used to construct the background state
-background_state_parameters = ( M² = 1e-8,       # s⁻², geostrophic shear squared
-                                 f = coriolis.f, # s⁻¹, Coriolis parameter
-                                 N = 1e-4,       # s⁻¹, buoyancy frequency
-                                Lz = grid.Lz )
+# Specify parameters that are used to construct the background state.
+background_state_parameters = ( M = 1e-4,       # s⁻¹, geostrophic shear
+                                f = coriolis.f, # s⁻¹, Coriolis parameter
+                                N = 1e-4,       # s⁻¹, buoyancy frequency
+                                H = grid.Lz )
 
-# Here, ``B`` is the background buoyancy field and ``V`` is the corresponding thermal wind
-V(x, y, z, t, p) = p.M² / p.f * (z - p.Lz/2)
-B(x, y, z, t, p) = p.M² * x + p.N^2 * (z - p.Lz/2)
+# Here, ``B`` is the background buoyancy field and ``V`` is the meridional velocity consistent with
[…]
