### 2023-07-15 044ddc1
more noise at center of domain
