### 2023-07-15 491775b
bit better thermal wind explanation
