### 2023-07-15 c794fbb
improved sinking flux to not make users use Oceananigans private API
