### 2023-07-15 59233cc
bit better end state
