+++ examples/eady.jl
@@ -20,7 +20,7 @@ using OceanBioME, Oceananigans, Printf
 using Oceananigans.Units
 using Random
 
-Random.seed!(1234)
+Random.seed!(11)
 
 # Construct a grid with uniform grid spacing.
 grid = RectilinearGrid(size = (32, 32, 8), extent = (1kilometer, 1kilometer, 100meters))
@@ -85,7 +85,7 @@ simulation = Simulation(model, Δt = 15minutes, stop_time = 10days)
 
 # Adapt the time step while keeping the CFL number fixed.
 wizard = TimeStepWizard(cfl = 0.75, diffusive_cfl = 0.75, max_Δt = 30minutes)
-simulation.callbacks[:wizard] = Callback(wizard, IterationInterval(1))
+simulation.callbacks[:wizard] = Callback(wizard, IterationInterval(5))
 nothing #hide
 
 # Create a progress message.
