+++ Project.toml
@@ -1,7 +1,7 @@
 name = "OceanBioME"
 uuid = "a49af516-9db8-4be4-be45-1dad61c5a376"
 authors = ["Jago Strong-Wright <js2430@damtp.cam.ac.uk>", "John R Taylor <j.r.taylor@damtp.cam.ac.uk>", "Si Chen <sc2350@damtp.cam.ac.uk>"]
-version = "0.3.0"
+version = "0.4.0"
 
 [deps]
 Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
@@ -18,7 +18,7 @@ KernelAbstractions = "0.9"
 Oceananigans = "0.83, 0.84"
 Roots = "2"
 StructArrays = "0.6.15"
-julia = "1.8"
+julia = "1.9"
 
 [extras]
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
