### 2023-07-15 8d52501
update Oceananigans compat entry
