### 2023-07-15 4e09e04
update Oceananigans compat entry
