+++ docs/Project.toml
@@ -16,4 +16,4 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 [compat]
 CairoMakie = "0.10"
 Literate = "≥2.9.0"
-Oceananigans = "0.83, 0.84"
+Oceananigans = "0.84.1"
