+++ Project.toml
@@ -15,7 +15,7 @@ StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
 Adapt = "3.4"
 JLD2 = "0.4.31"
 KernelAbstractions = "0.9"
-Oceananigans = "0.84.1"
+Oceananigans = "0.84.1, 0.85"
 Roots = "2"
 StructArrays = "0.6.15"
 julia = "1.9"
