+++ src/Boundaries/Sediments/simple_multi_G.jl
@@ -19,7 +19,7 @@ struct SimpleMultiG{FT, P1, P2, P3, P4, F, TE} <: FlatSediment
           refactory_fraction :: FT
 
     nitrate_oxidation_params :: P1
-       denitrifcation_params :: P2
+      denitrification_params :: P2
                anoxic_params :: P3
             solid_dep_params :: P4
 
@@ -37,7 +37,7 @@ end
                    slow_fraction::FT = 0.26,
                    refactory_fraction::FT = 0.1,
                    nitrate_oxidation_params::P1 = (A = - 1.9785, B = 0.2261, C = -0.0615, D = -0.0289, E = - 0.36109, F = - 0.0232),
-                   denitrifcation_params::P2 = (A = - 3.0790, B = 1.7509, C = 0.0593, D = - 0.1923, E = 0.0604, F = 0.0662),
+                   denitrification_params::P2 = (A = - 3.0790, B = 1.7509, C = 0.0593, D = - 0.1923, E = 0.0604, F = 0.0662),
                    anoxic_params::P3 = (A = - 3.9476, B = 2.6269, C = - 0.2426, D = -1.3349, E = 0.1826, F = - 0.0143),
                    depth = abs(znodes(grid, Face())[1]),
                    solid_dep_params::P4 = (A = 0.233, B = 0.336, C = 982, D = - 1.548, depth = depth))
@@ -84,12 +84,12 @@ function SimpleMultiG(; grid,
                                                        D = -0.0289,
                                                        E = - 0.36109,
                                                        F = - 0.0232),
-                       denitrifcation_params::P2 = (A = - 3.0790,
-                                                    B = 1.7509,
-                                                    C = 0.0593,
-                                                    D = - 0.1923,
-                                                    E = 0.0604,
-                                                    F = 0.0662),
+                       denitrification_params::P2 = (A = - 3.0790,
+                                                     B = 1.7509,
+                                                     C = 0.0593,
+                                                     D = - 0.1923,
+                                                     E = 0.0604,
+                                                     F = 0.0662),
                        anoxic_params::P3 = (A  = - 3.9476,
                                             B = 2.6269,
                                             C = - 0.2426,
@@ -120,7 +120,7 @@ function SimpleMultiG(; grid,
                                                    fast_redfield, slow_redfield,
                                                    fast_fraction, slow_fraction, refactory_fraction,
                                                    nitrate_oxidation_params,
-                                                   denitrifcation_params,
+                                                   denitrification_params,
                                                    anoxic_params,
                                                    solid_dep_params,
                                                    fields,
@@ -136,7 +136,7 @@ adapt_structure(to, sediment::SimpleMultiG) =
                  sediment.slow_fraction,
                  sediment.refactory_fraction,
                  sediment.nitrate_oxidation_params,
-                 sediment.denitrifcation_params,
+                 sediment.denitrification_params,
                  sediment.anoxic_params,
                  sediment.solid_dep_params,
                  adapt(to, sediment.fields),
@@ -193,13 +193,13 @@ sediment_fields(model::SimpleMultiG) = (C_slow = model.fields.C_slow,
                    sediment.nitrate_oxidation_params.D * log(k) * log(NH₄) +
                    sediment.nitrate_oxidation_params.E * log(Cᵐⁱⁿ * day) +
[…]
