+++ src/Models/AdvectedPopulations/NPZD.jl
@@ -344,5 +344,5 @@ adapt_structure(to, npzd::NPZD) =
                                                                    sinking_flux(i, j, grid, advection, Val(:P), bgc, tracers)
                  
 @inline carbon_flux(bgc::NPZD, tracers, i, j) = nitrogen_flux(bgc, tracers, i, j) * 6.56
-@inline remineralisation_receiver(::NPZD, tendenceies) = tendenceies.N
+@inline remineralisation_reciever(::NPZD) = :N
 end # module
