+++ docs/oceanbiome.bib
@@ -260,4 +260,15 @@
 	pages = {176-193},
 	title = {Model investigations of the North Atlantic spring bloom initiation},
 	volume = {138},
-	year = {2015}}
+	year = {2015}
+}
+
+
+@article{RemineralisationFraction,
+	author = {Dunne, John P. and Sarmiento, Jorge L. and Gnanadesikan, Anand},
+	journal = {Global Biogeochemical Cycles},
+	number = {4},
+	title = {A synthesis of global particle export from the surface ocean and cycling through the ocean interior and on the seafloor},
+	volume = {21},
+	year = {2007}
+}
