+++ src/Boundaries/Sediments/instant_remineralization.jl
@@ -92,5 +92,4 @@ sediment_fields(model::InstantRemineralisation) = (N_storage = model.fields.N_st
     @inbounds sediment.tendencies.Gⁿ.N_storage[i, j, 1] = burial_efficiency * flux
 
     @inbounds timestepper.Gⁿ[remineralisation_reciever(bgc)][i, j, 1] += flux * (1 - burial_efficiency) / Δz
-    end
 end
