### 2023-07-16 7997c70
better refs; use citet
