+++ src/Boundaries/Sediments/instant_remineralization.jl
@@ -5,7 +5,7 @@ Hold the parameters and fields the simplest benthic boundary layer where
 organic carbon is assumed to remineralise instantly with some portion 
 becoming N, and a fraction being perminantly burried.
 
-Burial efficiency from [RemineralisationFraction](@cite)
+Burial efficiency from [RemineralisationFraction](@citet).
 """
 struct InstantRemineralisation{FT, F, TE} <: FlatSediment
           burial_efficiency_constant1 :: FT
@@ -15,18 +15,16 @@ struct InstantRemineralisation{FT, F, TE} <: FlatSediment
                                fields :: F
                            tendencies :: TE
 
-    function InstantRemineralisation(burial_efficiency_constant1::FT,
-                                     burial_efficiency_constant2::FT,
-                                     burial_efficiency_half_saturaiton::FT,
-                                     fields::F,
-                                     tendencies::TE) where {FT, F, TE}
-
-        return new{FT, F, TE}(burial_efficiency_constant1,
-                              burial_efficiency_constant2,
-                              burial_efficiency_half_saturaiton,
-                              fields,
-                              tendencies)
-    end
+    InstantRemineralisation(burial_efficiency_constant1::FT,
+                            burial_efficiency_constant2::FT,
+                            burial_efficiency_half_saturaiton::FT,
+                            fields::F,
+                            tendencies::TE) where {FT, F, TE} =
+        new{FT, F, TE}(burial_efficiency_constant1,
+                        burial_efficiency_constant2,
+                        burial_efficiency_half_saturaiton,
+                        fields,
+                        tendencies)
 end
 
 """
