+++ docs/make.jl
@@ -12,7 +12,7 @@ CairoMakie.activate!(type = "svg")
 include("display_parameters.jl")
 
 bib_filepath = joinpath(dirname(@__FILE__), "oceanbiome.bib")
-bib = CitationBibliography(bib_filepath)
+bib = CitationBibliography(bib_filepath, style=:authoryear)
 
 # Examples
 
@@ -36,10 +36,10 @@ for example in example_scripts
 
     withenv("JULIA_DEBUG" => "Literate") do
         Literate.markdown(example_filepath, OUTPUT_DIR; 
-                          flavor = Literate.DocumenterFlavor(), 
-                          repo_root_url="https://oceanbiome.github.io/OceanBioME.jl", 
-                          execute=true,
-                          postprocess=replace_silly_warning)
+                          flavor = Literate.DocumenterFlavor(),
+                          repo_root_url = "https://oceanbiome.github.io/OceanBioME.jl",
+                          execute = false,
+                          postprocess = replace_silly_warning)
     end
 end
 
@@ -60,6 +60,7 @@ model_parameters = (LOBSTER(; grid = BoxModelGrid()),
                     GasExchange(; gas = :O₂).condition.parameters)
 
 gas_exchange_gas(::Val{G}) where G = G
+
 model_name(model) = if Base.typename(typeof(model)).wrapper == GasExchange
                         "$(gas_exchange_gas(model.gas)) air-sea exchange"
                     else
@@ -128,6 +129,7 @@ format = Documenter.HTML(
     prettyurls = get(ENV, "CI", nothing) == "true",
     canonical = "https://OceanBioME.github.io/OceanBioME/stable/",
     mathengine = MathJax3(),
+    assets = String["assets/citations.css"]
 )
 
 makedocs(bib,
@@ -136,7 +138,8 @@ makedocs(bib,
     format = format,
     pages = pages,
     modules = [OceanBioME],
-    doctest = true,
+    doctest = false,
+    draft = true,
     strict = true,
     clean = true,
     checkdocs = :exports
