+++ docs/src/model_components/sediment.md
@@ -1,6 +1,6 @@
 # [Sediment](@id sediment)
 
-We currently have one single-layer sediment model implemented. The model, proposed by [Soetaert2000](@cite), is a "G class" model that evolves carbon and nitrogen in three classes (fast, slow and refectory). The model is also only compatible with the LOBSTER biogeochemical model with carbonate chemistry, oxygen, and variable redfield options on. You also must ensure that the `open_bottom` option is on for particles to leave the bottom of the domain to the sediment model.
+We currently have one single-layer sediment model implemented. The model, proposed by [Soetaert2000](@citet), is a "G class" model that evolves carbon and nitrogen in three classes (fast, slow and refectory). The model is also only compatible with the LOBSTER biogeochemical model with carbonate chemistry, oxygen, and variable redfield options on. You also must ensure that the `open_bottom` option is on for particles to leave the bottom of the domain to the sediment model.
 
 It is straightforward to set up:
 
