+++ docs/src/model_components/biogeochemical/NPZ.md
@@ -1,3 +1,3 @@
 # [Nutrient Phytoplankton Zooplankton Detritus (NPZD) model](@id NPZD)
 
-The provided NPZD model is the low complexity model of [Kuhn2015](@cite).
+The provided NPZD model is the low complexity model of [Kuhn2015](@citet).

[…]
