### 2023-07-16 cdebd6d
convert gas exchange example to doctest
