### 2023-07-16 0f2cfcc
better doctest
