### 2023-07-16 1ff1606
convert doctest -> example
