+++ docs/src/model_components/air-sea-gas.md
@@ -6,39 +6,42 @@ Currently, the parameters for CO₂ and oxygen are included, but it would be ver
 
 It is straightforward to set up a boundary as an air-sea gas exchange:
 
-```jldoctest gasexchange
-julia> using OceanBioME
+```@setup gasexchange
+using OceanBioME
+CO₂_flux = GasExchange(; gas = :CO₂)
+using Oceananigans
 
-julia> CO₂_flux = GasExchange(; gas = :CO₂)
-FluxBoundaryCondition: ContinuousBoundaryFunction gasexchange_function at (Nothing, Nothing, Nothing)
+grid = RectilinearGrid(size=(3, 3, 30), extent=(10, 10, 200));
+
+model = NonhydrostaticModel(; grid,
+                              biogeochemistry = LOBSTER(; grid, carbonates = true),
+                              boundary_conditions = (DIC = FieldBoundaryConditions(top = CO₂_flux), ))
+```
+
+
+```@example gasexchange
+using OceanBioME
+
+CO₂_flux = GasExchange(; gas = :CO₂)
 ```
 
 Where the symbol specifies the exchanged gas (currently `:CO₂` or `:O₂`). This can then be passed in the setup of a BGC model, for example:
 
-```jldoctest gasexchange
-julia> using Oceananigans
+```@example gasexchange
+using Oceananigans
 
-julia> grid = RectilinearGrid(size=(3, 3, 30), extent=(10, 10, 200));
+grid = RectilinearGrid(size=(3, 3, 30), extent=(10, 10, 200));
 
-julia> model = NonhydrostaticModel(; grid,
-                                     biogeochemistry = LOBSTER(; grid, carbonates = true),
-                                     boundary_conditions = (DIC = FieldBoundaryConditions(top = CO₂_flux), ))
-NonhydrostaticModel{CPU, RectilinearGrid}(time = 0 seconds, iteration = 0)
-├── grid: 3×3×30 RectilinearGrid{Float64, Periodic, Periodic, Bounded} on CPU with 3×3×3 halo
-├── timestepper: QuasiAdamsBashforth2TimeStepper
-├── tracers: (NO₃, NH₄, P, Z, sPOM, bPOM, DOM, DIC, Alk)
-├── closure: Nothing
-├── buoyancy: Nothing
-└── coriolis: Nothing
+model = NonhydrostaticModel(; grid,
+                              biogeochemistry = LOBSTER(; grid, carbonates = true),
+                              boundary_conditions = (DIC = FieldBoundaryConditions(top = CO₂_flux), ))
 ```
 
 If the temperature and salinity are not included in the model they can be passed as functions
 (or even anonymous functions):
 
-```jldoctest gasexchange
-julia> T_function(x, y, z, t) = 12.0
-T_function (generic function with 1 method)
+```@example gasexchange
+T_function(x, y, z, t) = 12.0
 
-julia> CO₂_flux = GasExchange(; gas = :CO₂, temperature = T_function, salinity = (args...) -> 35)
-FluxBoundaryCondition: ContinuousBoundaryFunction gasexchange_function at (Nothing, Nothing, Nothing)
+CO₂_flux = GasExchange(; gas = :CO₂, temperature = T_function, salinity = (args...) -> 35)
 ```
