### 2023-07-16 d9dc58d
Added new model to docstring
