### 2023-07-19 #5601 @iuryt
I am happy to review this paper. The only thing though is that my Ph.D. defense will occur in August 1st. Thus, I will only be able to start the review after that, but as you mentioned that starting in the next 2 weeks is not a problem, you can count on me. Does that work? As I am very interested in the subject, I can try starting before that, if needed.
