### 2023-07-19 #5601 @francispoulin
> @tomchor @francispoulin @iuryt I'm looking for 2-3 reviewers for this submission to JOSS. Might you be interested in being a reviewer? As you probably know, reviews are conducted on github and are therefore open. We would hope to have the review wrapping up in 4-6 weeks total, meaning that the reviews would need to be started in the next 1-2 weeks since the process tends to be iterative between the reviewers and authors. Thank you for your consideration.

Hello, I would be happy to review this paper. Just let me know how you would like me to proceed.
