### 2023-07-19 #5601 @iuryt
> @iuryt Good luck on your upcoming defense! Do you think it would be realistic for you to start your review soon after your defense or will you have other tasks to work on? I do not think you should do a review before your defense! :)

I will probably have the day after off. So I can start by August 3rd.
