### 2023-07-19 #5669 @editorialbot
[REVIEW]: OceanBioME.jl: A flexible environment for modelling the coupled interactions between ocean biogeochemistry and physics
**Submitting author:** @jagoosw (<a href="http://orcid.org/0000-0002-7174-5283">Jago Strong-Wright</a>)
**Repository:** https://github.com/OceanBioME/OceanBioME.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.7.0
