+++ paper/paper.bib
@@ -132,8 +132,8 @@
 	pages = {759-776},
 	title = {Modelling seasonal growth and composition of the kelp Saccharina latissima},
 	volume = {24},
-	year = {2012}}
-
+	year = {2012}
+}
 
 @article{julia,
 	author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
@@ -143,4 +143,19 @@
 	title = {Julia: A fresh approach to numerical computing},
 	volume = {59},
 	year = {2017},
-   doi = {10.1137/141000671}}
+   doi = {10.1137/141000671}
+}
+
+@article{Friedlingstein2022,
+    title = {{Global Carbon Budget 2021}},
+    year = {2022},
+    journal = {Earth System Science Data},
+    author = {Friedlingstein, Pierre and Jones, Matthew W. and O'Sullivan, Michael and Andrew, Robbie M. and Bakker, Dorothee C.E. and Hauck, Judith and Le Qu{\'{e}}r{\'{e}}, Corinne and Peters, Glen P. and Peters, Wouter and Pongratz, Julia and Sitch, Stephen and Canadell, Josep G. and Ciais, Philippe and Jackson, Rob B. and Alin, Simone R. and Anthoni, Peter and Bates, Nicholas R. and Becker, Meike and Bellouin, Nicolas and Bopp, Laurent and Chau, Thi Tuyet Trang and Chevallier, Frédéric and Chini, Louise P. and Cronin, Margot and Currie, Kim I. and Decharme, Bertrand and Djeutchouang, Laique M. and Dou, Xinyu and Evans, Wiley and Feely, Richard A. and Feng, Liang and Gasser, Thomas and Gilfillan, Dennis and Gkritzalis, Thanos and Grassi, Giacomo and Gregor, Luke and Gruber, Nicolas and G{\"{u}}rses, Özgür and Harris, Ian and Houghton, Richard A. and Hurtt, George C. and Iida, Yosuke and Ilyina, Tatiana and Luijkx, Ingrid T. and Jain, Atul and Jones, Steve D. and Kato, Etsushi and Kennedy, Daniel and Goldewijk, Kees Klein and Knauer, Jürgen and Korsbakken, Jan Ivar and K{\"{o}}rtzinger, Arne and Landsch{\"{u}}tzer, Peter and Lauvset, Siv K. and Lef{\`{e}}vre, Nathalie and Lienert, Sebastian and Liu, Junjie and Marland, Gregg and McGuire, Patrick C. and Melton, Joe R. and Munro, David R. and Nabel, Julia E.M.S. and Nakaoka, Shin Ichiro and Niwa, Yosuke and Ono, Tsuneo and Pierrot, Denis and Poulter, Benjamin and Rehder, Gregor and Resplandy, Laure and Robertson, Eddy and R{\"{o}}denbeck, Christian and Rosan, Thais M. and Schwinger, Jörg and Schwingshackl, Clemens and S{\'{e}}f{\'{e}}rian, Roland and Sutton, Adrienne J. and Sweeney, Colm and Tanhua, Toste and Tans, Pieter P. and Tian, Hanqin and Tilbrook, Bronte and Tubiello, Francesco and Van Der Werf, Guido R. and Vuichard, Nicolas and Wada, Chisato and Wanninkhof, Rik and Watson, Andrew J. and Willis, David and Wiltshire, Andrew J. and Yuan, Wenping and Yue, Chao and Yue, Xu and Zaehle, Sönke and Zeng, Jiye},
+    number = {4},
+    month = {4},
+    pages = {1917--2005},
+    volume = {14},
+    publisher = {Copernicus GmbH},
+    doi = {10.5194/essd-14-1917-2022},
+    issn = {18663516}
+}
