+++ paper/paper.bib
@@ -159,3 +159,11 @@
     doi = {10.5194/essd-14-1917-2022},
     issn = {18663516}
 }
+
+@book{NationalAcademies2022,
+	address = {Washington, DC},
+	author = {National Academies of Sciences, Engineering, and Medicine},
+	publisher = {The National Academies Press},
+	title = {A Research Strategy for Ocean-based Carbon Dioxide Removal and Sequestration},
+	year = 2022
+}
