+++ paper/paper.bib
@@ -165,5 +165,6 @@
 	author = {National Academies of Sciences, Engineering, and Medicine},
 	publisher = {The National Academies Press},
 	title = {A Research Strategy for Ocean-based Carbon Dioxide Removal and Sequestration},
-	year = 2022
+	year = 2022,
+   doi = {10.17226/26278}
 }
