### 2023-07-21 99bfdbd
added near-global model figure
