### 2023-07-21 e2c0c71
Removed some repatition
