### 2023-07-21 dc7b0ef
minor edits to Statement of Need
