+++ paper/paper.bib
@@ -17,7 +17,7 @@
    journal = {Journal of Geophysical Research C: Oceans},
    month = {7},
    pages = {1-23},
-   title = {A four-dimensional mesoscale map of the spring bloom in the northeast Atlantic (POMME experiment): Results of a prognostic model},
+   title = {A four-dimensional mesoscale map of the spring bloom in the northeast {Atlantic} (POMME experiment): Results of a prognostic model},
    volume = {110},
    year = {2005},
 }
@@ -37,7 +37,6 @@
 @report{soetaert:2000,
    author = {Karline Soetaert and Jack J Middelburg and Peter M J Herman and Kerst Buis},
    journal = {Earth-Science Reviews},
-   keywords = {biogeochemical model,early diagenesis,nitrogen cycle,numerical model,oxygen cycle},
    pages = {173-201},
    title = {On the coupling of benthic and pelagic biogeochemical models},
    doi = {10.1016/S0012-8252(00)00004-0},
@@ -47,24 +46,22 @@
 }
 
 @misc{nemo,
-  author       = {{NEMO TOP Working Group}},
-  title        = {TOP – Tracers in Ocean Paradigm – The NEMO Tracers 
-                   engine},
-  month        = oct,
-  year         = 2018,
-  publisher    = {Zenodo},
-  version      = {v4.2.0},
-  doi          = {10.5281/zenodo.1471700},
-  url          = {https://doi.org/10.5281/zenodo.1471700}
+  author = {{NEMO TOP Working Group}},
+  title = {TOP – Tracers in Ocean Paradigm – The NEMO Tracers engine},
+  year = {2018},
+  publisher = {Zenodo},
+  version = {v4.2.0},
+  doi = {10.5281/zenodo.1471700},
+  url = {https://doi.org/10.5281/zenodo.1471700}
 }
 
 @article{macmods,
-    title={A Macroalgal Cultivation Modeling System (MACMODS): Evaluating the Role of Physical-Biological Coupling on Nutrients and Farm Yield},
-    volume={9},
-    DOI={10.3389/fmars.2022.752951},
-    journal={Frontiers in Marine Science},
-    author={Frieder, Christina A. and Yan, Chao and Chamecki, Marcelo and Dauhajre, Daniel and McWilliams, James C. and Infante, Javier and McPherson, Meredith L. and Kudela, Raphael M. and Kessouri, Fayçal and Sutula, Martha et al.},
-    year={2022}
+    title = {A Macroalgal Cultivation Modeling System (MACMODS): Evaluating the Role of Physical-Biological Coupling on Nutrients and Farm Yield},
+    volume = {9},
+    doi = {10.3389/fmars.2022.752951},
+    journal = {Frontiers in Marine Science},
+    author = {Frieder, Christina A. and Yan, Chao and Chamecki, Marcelo and Dauhajre, Daniel and McWilliams, James C. and Infante, Javier and McPherson, Meredith L. and Kudela, Raphael M. and Kessouri, Fayçal and Sutula, Martha et al.},
+    year = {2022}
 }
 
 @article{makie,
@@ -76,18 +73,17 @@
   number = {65},
   pages = {3349},
   author = {Simon Danisch and Julius Krumbiegel},
-  title = {Makie.jl: Flexible high-performance data visualization for Julia},
+  title = {Makie.jl: Flexible high-performance data visualization for {Julia}},
   journal = {Journal of Open Source Software}
 }
 
 @article{npzd,
[…]
