+++ paper/paper.bib
@@ -1,12 +1,12 @@
 @article{Oceananigans,
-    author = {Ramadhan, Ali and Wagner, Gregory LeClaire and Hill, Chris and Jean-Michel, Campin and Churavy, Valentin and Souza, Andre and Edelman, Adam and Ferrari, Raffaele and Marshall, John},
-    doi = {10.21105/joss.02018},
-    journal = {Journal of Open Source Software},
-    number = {53},
-    pages = {2018},
-    title = {{Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs}},
-    volume = {5},
-    year = {2020}
+   author = {Ramadhan, Ali and Wagner, Gregory LeClaire and Hill, Chris and Jean-Michel, Campin and Churavy, Valentin and Souza, Andre and Edelman, Adam and Ferrari, Raffaele and Marshall, John},
+   doi = {10.21105/joss.02018},
+   journal = {Journal of Open Source Software},
+   number = {53},
+   pages = {2018},
+   title = {{Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs}},
+   volume = {5},
+   year = {2020}
 }
 
 @article{lobster,
@@ -46,46 +46,46 @@
 }
 
 @misc{nemo,
-  author = {{NEMO TOP Working Group}},
-  title = {TOP – Tracers in Ocean Paradigm – The NEMO Tracers engine},
-  year = {2018},
-  publisher = {Zenodo},
-  version = {v4.2.0},
-  doi = {10.5281/zenodo.1471700},
-  url = {https://doi.org/10.5281/zenodo.1471700}
+   author = {{NEMO TOP Working Group}},
+   title = {TOP – Tracers in Ocean Paradigm – The NEMO Tracers engine},
+   year = {2018},
+   publisher = {Zenodo},
+   version = {v4.2.0},
+   doi = {10.5281/zenodo.1471700},
+   url = {https://doi.org/10.5281/zenodo.1471700}
 }
 
 @article{macmods,
-    title = {A Macroalgal Cultivation Modeling System (MACMODS): Evaluating the Role of Physical-Biological Coupling on Nutrients and Farm Yield},
-    volume = {9},
-    doi = {10.3389/fmars.2022.752951},
-    journal = {Frontiers in Marine Science},
-    author = {Frieder, Christina A. and Yan, Chao and Chamecki, Marcelo and Dauhajre, Daniel and McWilliams, James C. and Infante, Javier and McPherson, Meredith L. and Kudela, Raphael M. and Kessouri, Fayçal and Sutula, Martha et al.},
-    year = {2022}
+   title = {A Macroalgal Cultivation Modeling System (MACMODS): Evaluating the Role of Physical-Biological Coupling on Nutrients and Farm Yield},
+   volume = {9},
+   doi = {10.3389/fmars.2022.752951},
+   journal = {Frontiers in Marine Science},
+   author = {Frieder, Christina A. and Yan, Chao and Chamecki, Marcelo and Dauhajre, Daniel and McWilliams, James C. and Infante, Javier and McPherson, Meredith L. and Kudela, Raphael M. and Kessouri, Fayçal and Sutula, Martha et al.},
+   year = {2022}
 }
 
 @article{makie,
-  doi = {10.21105/joss.03349},
-  url = {https://doi.org/10.21105/joss.03349},
-  year = {2021},
-  publisher = {The Open Journal},
-  volume = {6},
-  number = {65},
-  pages = {3349},
-  author = {Simon Danisch and Julius Krumbiegel},
[…]
