+++ paper/paper.bib
@@ -115,7 +115,6 @@
    year = {2015}
 }
 
-
 @article{broch:2012,
    author = {Ole Jacob Broch and Dag Slagstad},
    doi = {10.1007/s10811-011-9695-y},
@@ -124,7 +123,7 @@
    journal = {Journal of Applied Phycology},
    month = {8},
    pages = {759-776},
-   title = {Modelling seasonal growth and composition of the kelp Saccharina latissima},
+   title = {Modelling seasonal growth and composition of the kelp {Saccharina} latissima},
    volume = {24},
    year = {2012}
 }
