### 2023-07-24 e413a27
Added equal contrib flag for alphabetically listed authors
Hope everyone is happy with this? Can discuss if not
