+++ paper/paper.md
@@ -13,14 +13,18 @@ authors:
     affiliation: "1, 2"
   - name: Si Chen
     affiliation: "1, 2"
+    equal-contrib: true
   - name: Navid C Constantinou
     orcid: 0000-0002-8149-4094
     affiliation: 3, 4
+    equal-contrib: true
   - name: Simone Silvestri
     affiliation: 5
+    equal-contrib: true
   - name: Gregory LeClaire Wagner
     orcid: 0000-0001-5317-2445
     affiliation: 5
+    equal-contrib: true
   - name: John R Taylor
     affiliation: "1, 2"
 affiliations:
