### 2023-08-02 #5669 @francispoulin
Another observation is that the codes often define units like minutes, hour, etc.  But Oceananigans.jl has a way to import these in one line.  Why not use that and reduce the number of lines in your code?
