### 2023-08-03 #5669 @navidcy
> Another observation is that the codes often define units like minutes, hour, etc. But Oceananigans.jl has a way to import these in one line. Why not use that and reduce the number of lines in your code?

I think that from one version and later Oceananigans dropped the definition of `year` as it was ambiguous (should a year be 365 days or 365.25 days or even more accurate..?) and that’s what brings the need or redefinition… I think… But that’s only in the examples, not in source code. I may be wrong. @jagoosw might have more to say.
