### 2023-08-03 #5669 @francispoulin
> Hi @francispoulin ,
>
> thanks for these comments!
>
> For now, I only want to raise one clarifying question regarding:
>
> > 1. The sample code that appears on the main page, and is then repeated  in the docs is a nice way to illustrate the power of the code but is slow to run. On my fairly new laptop, it took over 20 minutes.  I think that for a first example, you could decrease the resolution to have it run much faster, which would help the new user get started more easily.
>
> What was slow running the simulation or making the animation? If the latter, did you use CairoMakie or GLMakie?

Hello,

It was the simulation that was slow.

The animation looks great but if the run time can be reduced a bit I think it would give a more gentle introduction to the new users.
