### 2023-08-03 2592395
added more detail to 'tracer based' models
