### 2023-08-03 d380c64
Added more descriptive caption and panel labelling to eady example
