### 2023-08-03 6f661fd
attempt to fix table column width
