+++ paper/paper.md
@@ -129,16 +129,16 @@ Additionally, Strong-Wright (In prep.) is using the coupling of both the biogeoc
 
 | Example| OceanBioME features utilised        | Code location |
 |--------------|--------------|---------------|
-| Sub-mesoscale eddy | LOBSTER biogeochemical model<sup>1</sup> with carbonate model active, CO₂ exchange with the air<sup>2</sup>, Light attenuation<sup>3</sup>, mass conserving negativity protection<sup>4</sup> | `examples/eady.jl` with resolution increased to 64x64x16|
-| Near-global | NPZD model<sup>5</sup>, Light attenuation<sup>3</sup>| Work in progress, available upon request/for collaboration |
-| Idealised 1D model with kelp individuals |  LOBSTER biogeochemical model<sup>1</sup> with carbonate model and variable Redfield ratio for organic components active, CO₂ exchange with the air<sup>2</sup>, Light attenuation<sup>3</sup>, Saccharina Latissima (sugar kelp) model<sup>6</sup>, mass conserving negativity protection<sup>4</sup> | `paper/figures/column.jl`, similar to `examples/column.jl` and `examples/kelp.jl` |
+| Sub-mesoscale eddy | LOBSTER biogeochemical model$^1$ with carbonate model active, CO₂ exchange with the air$^2$, Light attenuation$^3$, mass conserving negativity protection$^4$ | `examples/eady.jl` with resolution increased to 64x64x16|
+| Near-global | NPZD model$^5$, Light attenuation$^3$| Work in progress, available upon request/for collaboration |
+| Idealised 1D model with kelp individuals |  LOBSTER biogeochemical model$^1$ with carbonate model and variable Redfield ratio for organic components active, CO₂ exchange with the air$^2$, Light attenuation$^3$, Saccharina Latissima (sugar kelp) model$^6$, mass conserving negativity protection$^4$ | `paper/figures/column.jl`, similar to `examples/column.jl` and `examples/kelp.jl` |
 
-<sup>1</sup> `LOBSTER`\
-<sup>2</sup> `GasExchange`\
-<sup>3</sup> `TwoBandPhotosyntheticallyActiveRadiation`\
-<sup>4</sup> `ScaleNegativeTracers`\
-<sup>5</sup> `NutrientPhytoplanktonZooplanktonDetritus`\
-<sup>6</sup> `SLatissima`
+$^1$ `LOBSTER`\
+$^2$ `GasExchange`\
+$^3$ `TwoBandPhotosyntheticallyActiveRadiation`\
+$^4$ `ScaleNegativeTracers`\
+$^5$ `NutrientPhytoplanktonZooplanktonDetritus`\
+$^6$ `SLatissima`
 
 <!---
 TODO: add eady plotting script
