### 2023-08-03 96d727d
worked, now for the subscripts
