### 2023-08-04 18537a6
updated figures scripts to latest version
