+++ paper/figures/column.jl
@@ -37,8 +37,7 @@ model = NonhydrostaticModel(; grid,
                               biogeochemistry = LOBSTER(; grid,
                                                           surface_phytosynthetically_active_radiation = PAR⁰,
                                                           carbonates = true,
-                                                          variable_redfield = true,
-                                                          advection_schemes = (sPOM = WENO(grid), bPOM = WENO(grid))),
+                                                          variable_redfield = true),
                               boundary_conditions = (DIC = FieldBoundaryConditions(top = CO₂_flux),
                                                      NO₃ = FieldBoundaryConditions(top = FluxBoundaryCondition(N_fixation))),
                               forcing = (NO₃ = N_upwelling, ),
@@ -98,8 +97,7 @@ model = NonhydrostaticModel(; grid,
                                                           surface_phytosynthetically_active_radiation = PAR⁰,
                                                           carbonates = true,
                                                           variable_redfield = true,
-                                                          particles,
-                                                          advection_schemes = (sPOM = WENO(grid), bPOM = WENO(grid))),
+                                                          particles),
                               boundary_conditions = (DIC = FieldBoundaryConditions(top = CO₂_flux),
                                                      NO₃ = FieldBoundaryConditions(top = FluxBoundaryCondition(N_fixation))),
                               forcing = (NO₃ = N_upwelling, ),
