### 2023-08-04 f129f68
improved concistency in formatting module and package names
