### 2023-08-04 406dbf9
improved examples table
