### 2023-08-04 d518b35
added model equations for bgc models
