+++ docs/src/model_components/biogeochemical/LOBSTER.md
@@ -21,3 +21,75 @@ Lodyc-DAMTP Ocean Biogeochemical Simulation Tools for Ecosystem and Resources (L
     ├── sPOM: 0.0 to -3.47e-5 m/s 
     └── bPOM: 0.0 to -0.0023148148148148147 m/s
 ```
+
+## Model equations
+
+### Core components
+When no additional components are activated the tracers $NO_3$, $NH_4$, $P$, $Z$, $sPOM$, $bPOM$, and $DOM$ evolve like:
+
+$\frac{\partial P}{\partial t} = (1-\gamma)\mu_P L_{PAR}\left(L_{NO_3} + L_{NH_4}\right)P - G_P - m_PP^2,$
+
+$\frac{\partial Z}{\partial t} = a_z\left(G_P + G_{sPOM}\right) - m_ZZ^2 - \mu_ZZ,$
+
+$\frac{\partial NO_3}{\partial t} = -\mu_PL_{PAR}L_{NO_3} + \mu_nNH_4,$
+
+$\frac{\partial NH_4}{\partial t} = -\mu_PL_{PAR}L_{NH_4} - \mu_nNH_4 
+     + \alpha_P\gamma\mu_P L_{PAR}\left(L_{NO_3} + L_{NH_4}\right)P
+     + \alpha_Z\mu_ZZ + \alpha_{d}\mu_{sPOM}sPOM + \alpha_{d}\mu_{bPOM}bPOM
+     + \mu_{DOM}DOM,$
+
+$\frac{\partial sPOM}{\partial t} = f_s[\left(1-a_Z)\left(G_P + G_{sPOM}\right) + m_PP^2 + m_ZZ^2\right] - G_{sPOM} - \mu_{sPOM}sPOM - \frac{\partial}{\partial z}(sPOM w_{s}),$
+
+$\frac{\partial bPOM}{\partial t} = (1-f_s)\left[(1-a_Z)\left(G_P + G_{sPOM}\right) + m_PP^2 + m_ZZ^2\right] - \mu_{bPOM}bPOM - \frac{\partial}{\partial z}(bPOM w_{b}),$
+
+$\frac{\partial DOM}{\partial t} = (1-\alpha_P)\gamma\mu_P L_{PAR}\left(L_{NO_3} + L_{NH_4}\right)R_PP + (1-\alpha_Z)\mu_ZR_ZZ + (1-\alpha_D)\mu_{sPOM}sPOM + (1-\alpha_D)\mu_{bPOM}bPOM - \mu_{DOM}DOM.$
+
+Where:
+
+$L_{PAR} = 1 - e^{-PAR/k_{PAR}},$
+
+$L_{NO_3} = \frac{NO_3}{NO_3 + k_{NO_3}}e^{-\psi NH_4},$
+
+$L_{NH_4} = \frac{NH_4}{NH_4 + k_{NH_4}},$
+
+$G_P = g_z\frac{\tilde{p}P}{k_z + \tilde{p}P + (1-\tilde{p})sPOM}Z,$
+
+$G_{sPOM} = g_z\frac{(1-\tilde{p})sPOM}{k_z + \tilde{p}P + (1-\tilde{p})sPOM}Z.$
+
+Additionally, the $sPOM$ and $bPOM$ detritus components sink with constant sinking speed.
+
+### Carbonate chemistry
+
+When the carbonate chemistry is activated additional tracers $DIC$ and $Alk$ evolve like:
+
[…]
