### 2023-08-04 8585a22
added gas exchange equations
