+++ docs/src/model_components/air-sea-gas.md
@@ -45,3 +45,89 @@ T_function(x, y, z, t) = 12.0
 
 CO₂_flux = GasExchange(; gas = :CO₂, temperature = T_function, salinity = (args...) -> 35)
 ```
+
+## Model equations
+
+### Gas transfer
+
+The gas flux is given by:
+
+$F = k(C_w - \alpha C_a),$
+
+where $C_w$ is the concentration in the water, $C_a$ the concentration in the air, $\alpha$ the Oswald solubility coefficient, and $k$ the gas transfer velocity. For carbon dioxide the flux is modified to:
+
+$F = k\beta\rho_o(pCO_{2w} - pCO_{2a}),$
+
+where $pCO_{2w}$ and $pCO_{2a}$ are the partial pressure of carbon dioxide in the water and air, $\beta$ is the Bunsen Solubility Coefficient, and $\rho_o$ is the density of the water.
+
+The gas transfer velocity is parameterised by the wind speed and Schmidt number, which in turn is parameterised by the temperature and salinity. The gas transfer velocity is given by:
+
+$k = 1.08\times10^{-6}u^2\left(\frac{Sc}{660}\right)^{-1/2},$
+
+where $u$ is the winds speed 10m above the surface, and Sc is the Schmidt number parameterised as:
+
+$Sc = A - BT + CT^2 - DT^3,$
+
+where $T$ is temperature in Kelvin and the other parameters are dependent on the gas type and given in [Parameters](@ref parameters).
+
+The solubilities are given by:
+
+$\alpha = 0.00367 T \exp{A_1 + 100\frac{A_2}{T} + A_3 \ln{\frac{T}{100}} + S\left(B_1 + \frac{B_2}{T} + \frac{B_3}{T^2}\right)},$
+
+and
+
+$\beta = \exp{A_1 + 100\frac{A_2}{T} + A_3 \ln{\frac{T}{100}} + S\left(B_1 + \frac{B_2}{T} + \frac{B_3}{T^2}\right)},$
+
+where $S$ is salinity in practical units and the other default parameters are given in [Parameters](@ref parameters).
+
+### Partial pressure of carbon dioxide
+
+We currently do not have the full OCMIP partial pressure formulation, instead we follow the simplified formulation where the partial pressure of CO$_2$ ($\mu$atm) in gas is found from:
+
+$pCO_{2a} = f_{CO_2}P_a,$
+
+where $f_{CO_2}$ is the air fraction (ppmv), and $P_a$ is the atmospheric pressure (atm). The ocean partial pressure ($\mu$atm) is derived from the dissolved inorganic carbon content, and alkalinity, as described by [tan1998](@citet), from the equilibrium of the following chemical system:
+
+$\ce{CO_2(g)} \ce{<=> CO_2(aq)},\ K_0=\frac{[\ce{O_2(aq)}]}{\ce{pCO_2}},$
+
+$\ce{CO_2(aq) + H_2O}\ce{<=> H^+ + HCO^-_3},\ K_1=\frac{\ce{[HCO^-_3][H^+]}}{\ce{[CO_2(aq)]}},$
+
+$\ce{HCO^-_3} \ce{<=> H^+ + CO^{2-}_3},\ K_2=\frac{\ce{[CO^{2-}_3][H^+]}}{\ce{HCO^-_3}}.$
+
[…]
