### 2023-08-04 221fb68
added light attenuation equations
