### 2023-08-04 dc2ea5d
Attempted to add a couple of sentences about the ease of implimenting and modifying models
