### 2023-08-04 de6ff8a
JOSS can't render unicode subscript 0
