### 2023-08-04 695b9f5
added kelp model equaions and fixed some links
