+++ docs/src/model_components/individuals/slatissima.md
@@ -6,3 +6,90 @@ The model tracks three variables, the frond area, A (dm²), carbon reserve, C (g
 
 Results could look something like this (from [StrongWright2022](@citet)):
 ![Example A, N, and C profiles from [StrongWright2022](@citet)](https://www.frontiersin.org/files/Articles/793977/fmars-08-793977-HTML/image_m/fmars-08-793977-g002.jpg)
+
+## Model equations
+
+As per [Broch2012](@citet) this model variables evolve as:
+
+$\frac{dA}{dt} = \left(\mu - \nu\right)A,$
+
+$\frac{dN}{dt} = J - \mu(N + N_\text{struct}),$
+
+$\frac{dC}{dt} = P(1 - E) - R - \mu(C + C_\text{struct}).$
+
+The apical frond loss given by:
+
+$\nu = \frac{10^{-6}\exp(\varepsilon A)}{1 + 10^{-6}(\exp(\varepsilon A) - 1)}.$
+
+The growth given by:
+
+$\mu = f_\text{area}f_\text{seasonal}f_\text{temp}\min\left(\mu_c, \max(\mu_{NO_3}, \mu_{NH_4})\right),$
+
+where:
+
+$f_\text{area} = m_1\exp(-(A/A_0)^2) + m_2,$
+
+$f_\text{seasonal} = a_1(1 + \text{sgn}(\lambda(n))|\lambda(n)|^{1/2}) + a_2,$
+
+$f_\text{temp} = \left{ \begin{array}{ll}
+                    0.08T + 0.2 & -1.8\leq T < 10 \\
+                    1           & 10 \leq T \leq 15 \\
+                    19/4 - T/4  & 15 < T \leq 19 \\
+                    0           & T > 19>
+                 \end{array} \right$
+
+where $n$ is the day of the year, $\lambda$ is the normalised day length change, and $T$ is the temperature in degrees centigrade. The limiting rates ($\mu_c$, $\mu_{NO_3}$, $\mu_{NH_4}$) depend on the availability of carbon giving:
+
+$\mu_c = 1 - \frac{C_\text{min}}{C},$
+
+and on the available nitrogen which is either limited by the instantanious uptake of ammonia, or the nitrogen reserve. To find these limits $J$, the nutrient uptake, must first be found [Fossberg2018](@citep). The uptake is calculated by first finding the $NO_3$ uptake rate:
+
+$J_{NO_3} = J_{NO_3\text{, max}}f_\text{curr}\frac{N_\text{max} - N}{N_\text{max} - N_\text{min}}\frac{NO_3}{k_{NO_3} + NO_3},$
+
+where $f_{curr} = c_1(1 - \exp(-u / c_2)) + c_3$ [Broch2019](@citep) where $u$ is the relative current speed. We then calculate the potential ammonia uptake:
+
+$\tilde{J}_{NH_4} = J_{NH_4\text{, max}}f_\text{curr}\frac{NH_4}{k_{NH_4} + NH_4}.$
+
+This results in a theoretical instantaneous area increase rate of:
+
+$\mu_{NH_4} = \frac{1}{N_\text{min} + N_\text{struct}}\frac{\tilde{J}_{NH_4}}{k_A},$
+
+or growth from reserves of:
+
+$\mu_{NO_3} = 1 - \frac{N_\text{min}}{N}.$
+
+The actual resulting ammonia uptake is then:
+
+$J_{NH_4} = \min\left(\tilde{J}_{NH_4}, \mu k_A (N_\text{min} + N_\text{struct})\right),$
+
+and the total uptake $J = \frac{J_{NH_4} + J_{NO_3}}{k_A}$.
+
[…]
