### 2023-08-04 61d6106
fixed slatissima model
