### 2023-08-04 cd9255d
reduced resolution of readme example, lets see how it looks
