### 2023-08-04 10ecf89
just read the docs and realised `$` was a depreciated way to write latex
