+++ docs/src/model_components/individuals/slatissima.md
@@ -11,84 +11,124 @@ Results could look something like this (from [StrongWright2022](@citet)):
 
 As per [Broch2012](@citet) this model variables evolve as:
 
-$\frac{dA}{dt} = \left(\mu - \nu\right)A,$
+```math
+\frac{dA}{dt} = \left(\mu - \nu\right)A,
+```
 
-$\frac{dN}{dt} = J - \mu(N + N_\text{struct}),$
+```math
+\frac{dN}{dt} = J - \mu(N + N_\text{struct}),
+```
 
-$\frac{dC}{dt} = P(1 - E) - R - \mu(C + C_\text{struct}).$
+```math
+\frac{dC}{dt} = P(1 - E) - R - \mu(C + C_\text{struct}).
+```
 
 The apical frond loss given by:
 
-$\nu = \frac{10^{-6}\exp(\varepsilon A)}{1 + 10^{-6}(\exp(\varepsilon A) - 1)}.$
+```math
+\nu = \frac{10^{-6}\exp(\varepsilon A)}{1 + 10^{-6}(\exp(\varepsilon A) - 1)}.
+```
 
 The growth given by:
 
-$\mu = f_\text{area}f_\text{seasonal}f_\text{temp}\min\left(\mu_c, \max(\mu_{NO_3}, \mu_{NH_4})\right),$
+```math
+\mu = f_\text{area}f_\text{seasonal}f_\text{temp}\min\left(\mu_c, \max(\mu_{NO_3}, \mu_{NH_4})\right),
+```
 
 where:
 
-$f_\text{area} = m_1\exp(-(A/A_0)^2) + m_2,$
+```math
+f_\text{area} = m_1\exp(-(A/A_0)^2) + m_2,
+```
 
-$f_\text{seasonal} = a_1(1 + \text{sgn}(\lambda(n))|\lambda(n)|^{1/2}) + a_2,$
+```math
+f_\text{seasonal} = a_1(1 + \text{sgn}(\lambda(n))|\lambda(n)|^{1/2}) + a_2,
+```
 
-$f_\text{temp} = \left[ \begin{array}{ll}
+```math
+f_\text{temp} = \left\{ \begin{array}{ll}
                     0.08T + 0.2 & -1.8\leq T < 10 \\
                     1           & 10 \leq T \leq 15 \\
                     19/4 - T/4  & 15 < T \leq 19 \\
                     0           & T > 19
-                 \end{array} \right]$
+                 \end{array} \right\}
+```
 
-where $n$ is the day of the year, $\lambda$ is the normalised day length change, and $T$ is the temperature in degrees centigrade. The limiting rates ($\mu_c$, $\mu_{NO_3}$, $\mu_{NH_4}$) depend on the availability of carbon giving:
+where ``n`` is the day of the year, ``\lambda`` is the normalised day length change, and ``T`` is the temperature in degrees centigrade. The limiting rates (``\mu_c``, ``\mu_{NO_3}``, ``\mu_{NH_4}``) depend on the availability of carbon giving:
 
-$\mu_c = 1 - \frac{C_\text{min}}{C},$
+```math
+\mu_c = 1 - \frac{C_\text{min}}{C},
+```
 
-and on the available nitrogen which is either limited by the instantanious uptake of ammonia, or the nitrogen reserve. To find these limits $J$, the nutrient uptake, must first be found [Fossberg2018](@citep). The uptake is calculated by first finding the $NO_3$ uptake rate:
[…]

[…]
