+++ paper/paper.md
@@ -93,7 +93,7 @@ For example, the `GasExchange` submodule calculates the carbon dioxide and oxyge
 We currently provide a simple Nutrient-Phytoplankton-Zooplankton-Detritus (NPZD) model [@npzd], and an intermediate complexity model, LOBSTER [@lobster], we have set up a straightforward "plug and play" framework to add additional tracers such as carbonate and oxygen chemistry systems and additional forcing. 
 A key feature of this package is the easy ability to modify models or add different formulations. 
 If a user wanted to implement a different model they could use the existing ones as a template and modify only a few lines of code where the ODEs are defined as functions.
-Alternatively, by harnessing another Julia feature, users could `import` specific functions from the models and overload only them.
+Alternatively, by harnessing another Julia feature, users could `import` specific functions from the models and redefine only them.
 For example, if a user wanted to modify the light limitation to phytoplankton growth in the NPZD model to copy that of the LOBSTER model they could pass different parameter values to the model and write:
 
 ```julia
