### 2023-08-04 #5669 @jagoosw
Hi @francispoulin

Thank you very much for your review! I’m glad that you think it’s useful.

I will try and respond to each point:

1. Thank you for pointing this out. I hadn’t realised that it does take quite a long time to run. When I run it on 8 threads (on an M1 Mac) it only takes a couple of minutes, but when I do it single-threaded as I imagine most new users might it does take a lot longer. I’ve made a PR (https://github.com/OceanBioME/OceanBioME.jl/pull/126) reducing the resolution and increasing the time step (as well as applying your other docs suggestions below).
2. This is a good point, I’ve tried adding some more description now.
3. Done
4. I have tried adding this at the end, is this what you meant? I also don’t have any public code for the near global model since it’s still very much an early work in progress but have put "available upon request".
5. I have also added to the PR a model equations section to the documentation, do you think the should be best added as an appendix section to the paper? We didn’t want to focus on specific model formulations here as we wanted to be fairly agnostic to particular models and planned on including details in any publications that use specific models. Also, the models are generally as previously published and we’ve linked their original publication where relevant. We also want to be careful to be clear that we didn’t come up with or test the model formulations (yet).
6. No, since the last few releases everything works with hydrostatic-free surface models too! We've also recently added this to the automated testing. I believe the only incompatibility we currently have is that the sediment models will not work properly with immersed boundaries, but this is a relatively simple fix that I intend to implement sometime soon.
7. That’s useful to know, thank you! I would be interested to see how much of a difference the positivity-preserving scheme makes vs the tendency for the biogeochemical forcing to go negative.
8. On the time units, I couldn’t see where we do this except for years in the examples? As @navidcy said we do this for years since it’s no longer defined by Oceananigans, and in these examples some of the forcing is cycling in idealised year length.
9. On the NetCDF issue, I’m not really sure what the best way to deal with it is. We have the same issue with CairoMakie, DataDeps, JLD2, Interpolations, and Statistics for one of the examples. Currently, we have them specified as `extras` in the Project file so for the CI they get installed, but I realise users (including me) don’t really know how to use this when running things locally.
10. I think this is a really good point and have added a few sentences about how users can modify/add models.

I will wait for the others to have a look before regenerating the pdf here (but it's available on the repo).

Thank you again, this was really helpful!
