### 2023-08-04 #5669 @navidcy
> 4. I have tried adding this at the end, is this what you meant? I also don’t have any public code for the near global model since it’s still very much an early work in progress but have put "available upon request".

@jagoosw can we avoid statements in the spirit of “available upon request”? I don’t like them at all because subconsciously I read them as “We don’t wanna share the code”
since my experience many time was that it was not available upon request after all. Perhaps a public repo with something, even if it evolves/changes in the future?
