+++ test/test_sediments.jl
@@ -9,6 +9,8 @@ using Oceananigans.Fields: TracerFields
 
 using Oceananigans.Operators: volume, Azᶠᶜᶜ
 
+import Oceananigans.Biogeochemistry: biogeochemical_drift_velocity
+
 function intercept_tendencies!(model, intercepted_tendencies)
     for tracer in keys(model.tracers)
         copyto!(intercepted_tendencies[tracer], model.timestepper.Gⁿ[tracer])
@@ -46,10 +48,17 @@ total_nitrogen(sed::InstantRemineralisation) = sum(sed.fields.N_storage)
 total_nitrogen(::LOBSTER, model) = sum(model.tracers.NO₃) + sum(model.tracers.NH₄) + sum(model.tracers.P) + sum(model.tracers.Z) + sum(model.tracers.DON) + sum(model.tracers.sPON) + sum(model.tracers.bPON)
 total_nitrogen(::NutrientPhytoplanktonZooplanktonDetritus, model) = sum(model.tracers.N) + sum(model.tracers.P) + sum(model.tracers.Z) + sum(model.tracers.D)
 
-function test_flat_sediment(grid, biogeochemistry; timestepper = :QuasiAdamsBashforth2)
-    model = NonhydrostaticModel(; grid, biogeochemistry, 
-                                  closure = nothing,
-                                  timestepper)
+function test_flat_sediment(grid, biogeochemistry, model; timestepper = :QuasiAdamsBashforth2)
+    model = isa(model, NonhydrostaticModel) ? model(; grid, 
+                                                      biogeochemistry, 
+                                                      closure = nothing,
+                                                      timestepper,
+                                                      buoyancy = nothing) :
+                                              model(; grid, 
+                                                      biogeochemistry, 
+                                                      closure = nothing,
+                                                      buoyancy = nothing,
+                                                      tracers = nothing)
 
     set_defaults!(model.biogeochemistry.sediment_model)
 
@@ -88,25 +97,65 @@ display_name(::LOBSTER) = "LOBSTER"
 display_name(::NutrientPhytoplanktonZooplanktonDetritus) = "NPZD"
 display_name(::SimpleMultiG) = "Multi-G"
 display_name(::InstantRemineralisation) = "Instant remineralisation"
+display_name(::RectilinearGrid) = "Rectilinear grid"
+display_name(::LatitudeLongitudeGrid) = "Latitude longitude grid"
+display_name(::ImmersedBoundaryGrid) = "Immersed boundary grid"
+display_name(::Type{NonhydrostaticModel}) = "Nonhydrostatic model"
+display_name(::Type{HydrostaticFreeSurfaceModel}) = "Hydrostatic free surface model"
 
-@testset "Sediment" begin
+
+bottom_height(x, y) = -1000 + 500 * exp(- (x^2 + y^2) / 250) # a hill
+
+@testset "Sediment integration" begin
     for architecture in (CPU(), )
-        grid = RectilinearGrid(architecture; size=(3, 3, 50), extent=(10, 10, 500))
-
-        for timestepper in (:QuasiAdamsBashforth2, :RungeKutta3),
-            sediment_model in (InstantRemineralisation(; grid), SimpleMultiG(; grid))
-            for biogeochemistry in (NutrientPhytoplanktonZooplanktonDetritus(; grid, open_bottom = true,
-                                                                               sediment_model),
-                                    LOBSTER(; grid,
-                                              carbonates = true, oxygen = true, variable_redfield = true, 
-                                              open_bottom = true,
-                                              sediment_model))
-
[…]
