+++ src/Models/AdvectedPopulations/LOBSTER/LOBSTER.jl
@@ -321,8 +321,9 @@ function LOBSTER(; grid,
 
                    particles::P = nothing) where {FT, LA, S, P}
 
-    (!isnothing(sediment_model) && open_bottom) ||
+    if !isnothing(sediment_model) && !open_bottom
         @warn "You have specified a sediment model but not `open_bottom` which will not work as the tracer will settle in the bottom cell"
+    end
 
     sinking_velocities = setup_velocity_fields(sinking_speeds, grid, open_bottom)
 
