### 2023-08-06 598405b
Added near global example code
