+++ paper/paper.md
@@ -143,7 +143,7 @@ Additionally, Strong-Wright (In prep.) is using the coupling of both the biogeoc
 | Example| OceanBioME features utilised        | Code location |
 |--------------|--------------|---------------|
 | Sub-mesoscale eddy (\autoref{eady}) | LOBSTER biogeochemical model$^1$ with carbonate model active, CO$_2$ exchange with the air$^2$, Light attenuation$^3$, mass conserving negativity protection$^4$ | `examples/eady.jl` with resolution increased to 64x64x16|
-| Near-global (\autoref{global}) | Light attenuation$^3$, NPZD model$^5$| Work in progress, available upon request/for collaboration |
+| Near-global (\autoref{global}) proof of concept | Light attenuation$^3$, NPZD model$^5$| [https://github.com/OceanBioME/GlobalOceanBioME.jl/releases/tag/v0.0.1](https://github.com/OceanBioME/GlobalOceanBioME.jl/releases/tag/v0.0.1) |
 | Idealised 1D model with kelp individuals (\autoref{column}) |  LOBSTER biogeochemical model$^1$ with carbonate model and variable Redfield ratio for organic components active, CO$_2$ exchange with the air$^2$, light attenuation$^3$, mass conserving negativity protection$^4$, and Saccharina Latissima (sugar kelp) model$^6$ | `paper/figures/column.jl`, similar to `examples/column.jl` and `examples/kelp.jl` |
 
 $^1$ `LOBSTER`\
