### 2023-08-06 3f81502
increased example resolution
