+++ docs/src/index.md
@@ -31,7 +31,7 @@ As a simple example let's run a Nutrient-Phytoplankton-Zooplankton-Detritus (NPZ
 using OceanBioME
 using Oceananigans, Oceananigans.Units
 
-grid = RectilinearGrid(CPU(), size = (128, 16), extent = (500meters, 100meters), topology = (Bounded, Flat, Bounded))
+grid = RectilinearGrid(CPU(), size = (160, 32), extent = (500meters, 100meters), topology = (Bounded, Flat, Bounded))
 
 biogeochemistry = NutrientPhytoplanktonZooplanktonDetritus(; grid, open_bottom = true)
 
