### 2023-08-06 #5669 @iuryt
I relate to some of the comments that @francispoulin mentioned, but I will put it here anyway.

For now, I am only reviewing the document and giving my comments based on it. I will add other post for more comments regarding the usability and documentation.
I am really impressed with how well written and modular this package is. This is what I like most from Oceananigans.jl community. The extensive knowledge of the physical and biological dynamics made you to setup the code in a way that users can combine and make it adaptable for a myriad of problems. This is the major good point I have about this paper and module.

## General comments

1) A better description of NPZD models is needed for the documentation.
2) I really liked that you made it possible to run as a box model. This is very important for testing, as you mentioned, but I can see many applications for education purposes as well. You should mention that in the text.

## Figures

**Figure 1:** I like this figure and I think it shows how powerful the package is, but I have some comments regarding the plots. If the tick labels are not important, you can just remove them, but if you think you should keep them, then I think it is important to enlarge the font size.

This is very neat picking, but there are two glitch-like things that I am not sure are related to the plotting package or something on the output. I believe it is from the plotting package, but just wanted to make sure.
![image](https://github.com/openjournals/joss-reviews/assets/5797727/830d8e01-c954-4743-8558-af6473bb707c)

Maybe labelling the panels makes it easier to reference in the text.

![image](https://github.com/openjournals/joss-reviews/assets/5797727/0ea38079-1de1-47c2-8859-bc3e6a46c7da)

**Figure 3:**

- Remove "Here we show the.." and just start with "Results of a 1D model,..."
- I understand that it might not look nice, but I would be consistent with the direction and use of the colorbars for different properties. The colorbar used for Phytoplankton changes direction from Figure 1 to Figure 2 to Figure 3. I would use a different colorbar for Nutrients in Figure 3 (maybe the same you used in Figure 1?)
- Add labels and explain each panel in the Figure caption.
- This is a personal preference and I don't think you should follow that if you don't prefer, but I would limit the number of colors in the colorbar for some plots. Perceiving a continuous gradient is hard, but if you have a discrete number of colors it is more easy to notice and relate the colors to the range of values. For instance, in Figure 3, you could have a color for each 0.1 mmol N/ m3 interval for the nutrient concentration.
