### 2023-08-06 #5669 @iuryt
I am thinking if it is better to add Oceananigans.jl as a requirement as you clearly say that OceanBioMe was built to be coupled with Oceananigans.jl, but I understand that you can run this package without using Oceananigans.jl. This is actually a very cool feature I found for this package.

The problem is that the first example in the docs uses Oceananigans.jl and if the user simply try the quick install the way it is right now, this is going to give `ArgumentError: Package Oceananigans not found in current path.`.

One alternative would be to give a simulation without Oceananigans.jl as the first example (box model?). Another alternative would be to add a code snippet on how to install Oceananigans.jl, refer to their webpage or mention in the beginning of the section for "Running your first model" that this simulation requires Oceananigans.jl.

Is that a way to give two set of requirements? Maybe having a `minimal` (which does not include Oceananigans.jl) and a full (which includes Oceananigans.jl and all other packages used in the tutorials). Then you could mention the two options in the quick install and mention in the beginning of the "Running your first model" section that this simulation uses the full instalation.
