### 2023-08-06 #5669 @iuryt
For the documentation. I mentioned in the first comment that it is better to describe equations and variables for the NPZD model, but I also think that the documentation for the LOBSTER model can be improved.

I think it is better to have a description similar to the one from the Appendix of https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2004JC002588
Including the equations and how the parameters are named in the code.

I am a fan of flow charts and for biogeochemical models this is a nice way to represent the equations graphically. I spent some time creating a version of figure A1 of the reference that you could use for the documentation. I provide an excalidraw link for this chart that you can use to create something similar for the NPZD model.

![LOBSTER_light](https://github.com/openjournals/joss-reviews/assets/5797727/41cedb44-1d00-4cb6-8f52-100afdf218da)

![LOBSTER_dark](https://github.com/openjournals/joss-reviews/assets/5797727/94766ac8-d394-4e9d-9d9d-0e57f312db15)

https://excalidraw.com/#json=xzcwARQHhrI2VN1P0QWKd,gOn5OqXs1gDr1Sfhe8VLCw
