### 2023-08-07 #5669 @iuryt
By the way, one thing I loved about this package is that the parameters has a name that has a meaning, which makes the code easier to read, understand and use. In other words, instead of calling the mortality by some Greek letter, you simply call it `phytoplankton_mortality`.
