### 2023-08-07 b9fa26f
Added parameter variable names for NPZD
