+++ docs/src/model_components/biogeochemical/NPZ.md
@@ -24,6 +24,21 @@ Here ``\mu_\text{max} = \mu_0Q_{10}(T)`` and ``l_{XY} = l_{XY0} Q_{10}(T)`` wher
 
 Additionally, the phytoplankton and detritus sink at a constant rate.
 
+### Parameter variable names
+
+| Symbol             | Variable name                  | Units            |
+|--------------------|--------------------------------|------------------|
+| ``\alpha``         | `initial_photosynthetic_slope` | 1 / (W / m² / s) |
+| ``\mu_\text{max}`` | `base_maximum_growth`          | 1 / s            |
+| ``k_N``            | `nutrient_half_saturation`     | mmol N / m³      |
+| ``l_{PN}``         | `base_respiration_rate`        | 1 / s            |
+| ``l_{PD}``         | `phyto_base_mortality_rate`    | 1 / s            |
+| ``g_\text{max}``   | `maximum_grazing_rate`         | 1 / s            |
+| ``k_P``            | `grazing_half_saturation`      | mmol N / m³      |
+| ``\beta``          | `assimulation_efficiency`      | -                |
+| ``l_{ZD}``         | `zoo_base_mortality_rate`      | 1 / s            |
+| ``r_{DN}``         | `remineralization_rate`        | 1 / s            |
+
 All default parameter values are given in [Parameters](@ref parameters).
 
 ## Model conservation
