+++ docs/src/model_components/biogeochemical/LOBSTER.md
@@ -124,7 +124,40 @@ In this case the organic components are split into nitrogen and carbon compartme
 
 Additionally, the ``DIC`` and ``Alk`` equations are modified to replace each ``XOM \cdot R_O`` with the corresponding ``XOC``.
 
-All default parameter values are given in [Parameters](@ref parameters); and a more thorough explanation of new terms will be included in a publication that is in prep, or is available upon request.
+### Parameter variable names
+
+| Symbol             | Variable name                          | Units              |
+|--------------------|----------------------------------------|--------------------|
+| ``\tilde{p}``      | `phytoplankton_preference`             | -                  |
+| ``g_z``            | `maximum_grazing_rate`                 | 1 / s              |
+| ``k_z``            | `grazing_half_saturation`              | mmol N / m³        |
+| ``k_{PAR}``        | `light_half_saturation`                | W / m²             |
+| ``\psi``           | `nitrate_ammonia_inhibition`           | -                  |
+| ``k_{NO_3}``       | `nitrate_half_saturation`              | mmol N / m³        |
+| ``k_{NH_4}``       | `ammonia_half_saturation`              | mmol N / m³        |
+| ``\mu_P``          | `maximum_phytoplankton_growthrate`     | 1 / s              |
+| ``a_z``            | `zooplankton_assimilation_fraction`    | -                  |
+| ``m_Z``            | `zooplankton_mortality`                | 1 / s / mmol N/m³  |
+| ``\mu_z``          | `zooplankton_excretion_rate`           | 1 / s              |
+| ``m_P``            | `phytoplankton_mortality`              | 1 / s              |
+| ``\mu_{sPOM}``     | `small_detritus_remineralisation_rate` | 1 / s              |
+| ``\mu_{bPOM}``     | `large_detritus_remineralisation_rate` | 1 / s              |
+| ``\gamma``         | `phytoplankton_exudation_fraction`     | -                  |
+| ``\mu_n``          | `nitrifcaiton_rate`                    | 1 / 2              |
+| ``\alpha_P``       | `ammonia_fraction_of_exudate`          | -                  |
+| ``\alpha_Z``       | `ammonia_fraction_of_excriment`        | -                  |
+| ``\alpha_d``       | `ammonia_fraction_of_detritus`         | -                  |
+| ``R_P``            | `phytoplankton_redfield`               | mmol C / mmol N    |
+| ``R_O``            | `organic_redfield`                     | mmol C / mmol N    |
+| ``R_{Chl:N}``      | `phytoplankton_chlorophyll_ratio`      | mg Chl / mmol N    |
+| ``\rho_{CaCO_3}``  | `organic_carbon_calcate_ratio`         | mmol CaCO₃/ mmol C |
+| ``R_{O_2}``        | `respiraiton_oxygen_nitrogen_ratio`    | mmol O / mmol N    |
+| ``R_{nit}``        | `nitrifcation_oxygen_nitrogen_ratio`   | mmol O / mmol N    |
+| ``f_s``            | `slow_sinking_mortality_fraction`      | -                  |
+| ``\mu_{DOM}``      | `disolved_organic_breakdown_rate`      | 1 / s              |
+| ``\eta``           | `zooplankton_calcite_dissolution`      | -                  |
+
+All default parameter values are given in [Parameters](@ref parameters); and a more thorough explanation of new terms will be included in a publication that is in prep.
 
 ## Model conservations
 
