+++ docs/src/model_components/individuals/slatissima.md
@@ -101,19 +101,19 @@ P = P_S\left[1 - \exp\left(\frac{\alpha PAR}{P_S}\right)\right]\exp\left(\frac{\
 where ``PAR`` is the photosynthetically available radiation and
 
 ```math
-P_S = \frac{\alpha PAR}{\ln(1 + \alpha/\beta)}.
+P_S = \frac{\alpha I_\text{sat}}{\ln(1 + \alpha/\beta)}.
 ```
 
 ``\alpha`` is a constant but $\beta$ depends on the maximum photosynthetic rate which is defined by both:
 
 ```math
-P_\text{max} = \frac{\alpha PAR}{\ln(1 + \alpha/\beta)}\left(\frac{\alpha}{\alpha + \beta}\right)\left(\frac{\beta}{\alpha + \beta}\right) ^ {\beta/\alpha},
+P_\text{max} = \frac{\alpha I_\text{sat}}{\ln(1 + \alpha/\beta)}\left(\frac{\alpha}{\alpha + \beta}\right)\left(\frac{\beta}{\alpha + \beta}\right) ^ {\beta/\alpha},
 ```
 
 and
 
 ```math
-P_\text{max} = \frac{P_1\exp\left(\frac{T_{AP}}{T_{P1}} - \frac{T_{AP}}{T}\right)}{1 + \exp\left(\frac{T_{APL}}{T} - \frac{T_{APL}}{PL}\right) + \exp\left(\frac{T_{APH}}{T} - \frac{T_{APH}}{PH}\right)},
+P_\text{max} = \frac{P_1\exp\left(\frac{T_{AP}}{T_{P1}} - \frac{T_{AP}}{T}\right)}{1 + \exp\left(\frac{T_{APL}}{T} - \frac{T_{APL}}{T_{PL}}\right) + \exp\left(\frac{T_{APH}}{T} - \frac{T_{APH}}{T_{PH}}\right)},
 ```
 
 where ``T`` is the temperature in kelvin and the ``T_X`` are Arrhenius temperature constants. We solve these iteratively to find ``\beta``.
@@ -130,6 +130,67 @@ As per [Broch2013](@citet) ``R``, the respiration rate, is given by:
 R = \left[R_A\left(\frac{\mu}{\mu_\text{max}} + \frac{J}{J_\text{max}}\right) + R_B\right]\exp\left(\frac{T_{ARR}}{T_1} - \frac{T_ARR}{T}\right).
 ```
 
-All other undefined symbols are constants which can be found in [Parameters](@ref parameters)
+### Parameter variable names
+
+| Symbol                   | Variable name                        | Units                   |
+|--------------------------|--------------------------------------|-------------------------|
+| ``A_0``                  | `growth_rate_adjustement`            | 1 / dm²                 |
+| ``\alpha``               | `photosynthetic_efficiency`          | gC / dm² / s / einstein |
+| ``C_\text{min}``         | `minimum_carbon_reserve`             | gC / gSW                |
+| ``C_\text{struct}``      | `structural_carbon`                  | gC / gSW                |
+| ``\gamma``               | `exudation`                          | gC / g                  |
+| ``\varepsilon``          | `erosion`                            | 1 / dm²                 |
+| ``I_\text{sat}``         | `saturation_irradiance`              | einstein                |
+| ``k_A``                  | `structural_dry_weight_per_area`     | g / dm²                 |
+| ``k_dw``                 | `structural_dry_to_wet_weight`       | -                       |
+| ``k_C``                  | `carbon_reserve_per_carbon`          | g / gC                  |
+| ``k_N``                  | `nitrogen_reserve_per_nitrogen`      | g / gN                  |
+| ``N_\text{min}``         | `minimum_nitrogen_reserve`           | gN / gSW                |
[…]
