+++ docs/src/model_components/biogeochemical/NPZ.md
@@ -2,6 +2,8 @@
 
 The provided NPZD model is the low complexity model of [Kuhn2015](@citet).
 
+![Diagram of NPZD formulation](assets/npzd.svg)
+
 ## Model equations
 
 ```math
