### 2023-08-07 e912a22
Change buoyancy front animation in docs to gif so can be used in readme
