+++ docs/src/model_components/biogeochemical/NPZ.md
@@ -2,7 +2,7 @@
 
 The provided NPZD model is the low complexity model of [Kuhn2015](@citet).
 
-![Diagram of NPZD formulation](assets/npzd.svg)
+![Diagram of NPZD formulation](npzd.svg)
 
 ## Model equations
 
