### 2023-08-07 ead737d
added better fig 3 description and labelling + other review suggestions
