### 2023-08-07 #5669 @jagoosw
Hi @iuryt,

Thank you very much for this thorough review and praise!

Hopefully we have addressed the problems with the description of the models in this pr https://github.com/OceanBioME/OceanBioME.jl/pull/126 which we will probably merge in the next few days. We’ve added model equations and parameter lists, and I have just added diagrams as you suggested (thank you for this idea!).

I am not sure that the colorers are necessary, but do think we need some kind of indication as to which direction the colours go in (for each one the darker colours represent higher values incase we wanted to remove them). I will have a play to see how it looks good. And yes, the glitches are from Makie.

I have flipped the colours in Figure 2 (thank you for spotting this, this way around is definitely more appropriate), and have added panel letters and description. I’m going to have a play with all of the figures and see if discrete colours look better here.

> I am thinking if it is better to add Oceananigans.jl as a requirement as you clearly say that OceanBioMe was built to be coupled with Oceananigans.jl, but I understand that you can run this package without using Oceananigans.jl. This is actually a very cool feature I found for this package.

I’m not sure I understand what you mean by this. Oceananigans is currently a dependency even if a use only wants to run box models (although I don’t think this is ideal, we went for this option as I think it would be a lot of work to get stand alone box models). If you `] add OceanBioME` all of the examples should work (``\pm`` problems mentioned above with examples only dependencies e.g. Makie). In the examples we have an “install dependance” line as this seems standard practice in Julia docs examples.
