### 2023-08-07 #5669 @iuryt
To reproduce the error, I did the following.

Created a new directory and an empty `Project.toml`
```bash
mkdir testing && cd testing && touch Project.toml
```

Install `OceanBioMe` on the Julia project
```bash
julia --project -e 'using Pkg; Pkg.add("OceanBioME")'
```

Now I simply try
```bash
julia --project -e 'using Oceananigans'
```
which returns the following error
```julia
ERROR: ArgumentError: Package Oceananigans not found in current path.
- Run `import Pkg; Pkg.add("Oceananigans")` to install the Oceananigans package.
Stacktrace:
 [1] macro expansion
   @ ./loading.jl:1630 [inlined]
 [2] macro expansion
   @ ./lock.jl:267 [inlined]
 [3] require(into::Module, mod::Symbol)
   @ Base ./loading.jl:1611
```
