### 2023-08-07 #5669 @jagoosw
Oh I see. I don't think it's possible for one package to make another available in a new project that it's installed in, e.g. Oceananigans depends on KernelAbstractions but if you do the above you can't `using KernelAbstractions` without adding it first.

When you add Oceananigans to the above project it doesn't actually have to install it either it just modifies the Project to make it available on the top level.
```julia
julia> using Oceananigans
 │ Package Oceananigans not found, but a package named Oceananigans is available from a registry.
 │ Install package?
 │   (testing) pkg> add Oceananigans
 └ (y/n/o) [y]:
   Resolving package versions...
    Updating `~/tmp/testing/Project.toml`
⌅ [9e8cae18] + Oceananigans v0.85.0
  No Changes to `~/tmp/testing/Manifest.toml`
```

Perhaps we should move the box model to the top example, and move the `JLD2` and `CairoMakie` dependency (for loading and plotting the results) lower, so that off the bat you just have to have OceanBioME installed.
