### 2023-08-07 45f9583
rearranged examples
