### 2023-08-08 #5669 @iuryt
Yes. Another alternative would be to add a code cell that adds the dependencies. I particularly like the lock exchange example. You can choose.
