### 2023-08-08 #5669 @navidcy
> Perhaps we should move the box model to the top example, and move the `JLD2` and `CairoMakie` dependency (for loading and plotting the results) lower, so that off the bat you just have to have OceanBioME installed.

Personally I like that the first example is a fluids example and not a box model.
