### 2023-08-08 4c13967
Added lower res gif
