### 2023-08-08 #5669 @jagoosw
Oh I see. Yes, thats a good point!

We have to define `T` for the NPZD model, but the proper way would be to use a T and S-dependant buoyancy model.
