### 2023-08-08 778889e
added Oceananigans install for quickstart
