+++ README.md
@@ -28,9 +28,11 @@ julia> Pkg.add("OceanBioME")
 ```
 
 ## Running your first model
-As a simple example lets run a Nutrient-Phytoplankton-Zooplankton-Detritus (NPZD) model in a two-dimensional simulation of a buoyancy front:
+As a simple example lets run a Nutrient-Phytoplankton-Zooplankton-Detritus (NPZD) model in a two-dimensional simulation of a buoyancy front. This also requires Oceananigans, so we will add that first:
 
 ```julia
+using Pkg; Pkg.add("Oceananigans")
+
 using OceanBioME, Oceananigans
 using Oceananigans.Units
 
