+++ docs/src/index.md
@@ -25,33 +25,39 @@ julia> Pkg.add("OceanBioME")
 
 ## Running your first model
 
-As a simple example lets run a Nutrient-Phytoplankton-Zooplankton-Detritus (NPZD) model in a two-dimensional simulation of a buoyancy front. This also requires Oceananigans so we will add that first:
+As a simple example lets run a Nutrient-Phytoplankton-Zooplankton-Detritus (NPZD) model in a two-dimensional simulation of a buoyancy front. This also requires Oceananigans, so we will add that first:
 
 ```julia
 using Pkg; Pkg.add("Oceananigans")
 
-using OceanBioME
-using Oceananigans, Oceananigans.Units
+using OceanBioME, Oceananigans
+using Oceananigans.Units
 
-grid = RectilinearGrid(CPU(), size = (160, 32), extent = (500meters, 100meters), topology = (Bounded, Flat, Bounded))
+grid = RectilinearGrid(CPU(), size = (160, 32), extent = (500meters, 
+100meters), topology = (Bounded, Flat, Bounded))
 
-biogeochemistry = NutrientPhytoplanktonZooplanktonDetritus(; grid, open_bottom = true)
+biogeochemistry = NutrientPhytoplanktonZooplanktonDetritus(; grid, 
+open_bottom = true)
 
 model = NonhydrostaticModel(; grid, biogeochemistry,
-                              buoyancy = BuoyancyTracer(), tracers = :b,
                               advection = WENO(; grid),
+			                  buoyancy = SeawaterBuoyancy(constant_salinity = true),
                               closure = AnisotropicMinimumDissipation())
 
-bᵢ(x, y, z) = ifelse(x < grid.Lx/2, 1e-4, 1e-3)
+Tᵢ(x, y, z) = ifelse(x < grid.Lx/2, 1, 10)
 
-set!(model, b = bᵢ, N = 5.0, P = 0.1, Z = 0.1, T = 18.0)
+set!(model, N = 5.0, P = 0.1, Z = 0.1, T = Tᵢ)
 
-simulation = Simulation(model; Δt = 4.0, stop_time = 3hours)
+simulation = Simulation(model; Δt = 1.0, stop_time = 3hours)
 
-simulation.output_writers[:tracers] = JLD2OutputWriter(model, model.tracers,
-                                                       filename = "buoyancy_front.jld2",
-                                                       schedule = TimeInterval(1minute),
-                                                       overwrite_existing = true)
+simulation.output_writers[:tracers] = JLD2OutputWriter(model, 
+model.tracers,
+                                                       filename = 
+"buoyancy_front.jld2",
+                                                       schedule = 
+TimeInterval(1minute),
+                                                       overwrite_existing 
+= true)
 
 run!(simulation)
 ```
@@ -59,22 +65,22 @@ run!(simulation)
 We can then load the saved output and visualize it:
 
 ```@example quickstart
-b = FieldTimeSeries("buoyancy_front.jld2", "b")
+T = FieldTimeSeries("buoyancy_front.jld2", "T")
 P = FieldTimeSeries("buoyancy_front.jld2", "P")
 
-xb, yb, zb = nodes(b)
+xT, yT, zT = nodes(T)
 xP, yP, zP = nodes(P)
 
-times = b.times
+times = T.times
 
 using CairoMakie
 
 n = Observable(1)
[…]
