+++ docs/src/index.md
@@ -25,7 +25,7 @@ julia> Pkg.add("OceanBioME")
 
 ## Running your first model
 
-As a simple example lets run a Nutrient-Phytoplankton-Zooplankton-Detritus (NPZD) model in a two-dimensional simulation of a buoyancy front. This also requires Oceananigans, so we will add that first:
+As a simple example lets run a Nutrient-Phytoplankton-Zooplankton-Detritus (NPZD) model in a two-dimensional simulation of a buoyancy front. This example requires Oceananigans, so we install that first:
 
 ```julia
 using Pkg; Pkg.add("Oceananigans")
