### 2023-08-08 a997dcd
reduced temp gradient
