### 2023-08-08 ef3601f
added back `@example`
