### 2023-08-09 150d55c
adjusted readme example (takes about 1.5min to run now)
