+++ docs/src/index.md
@@ -33,31 +33,29 @@ using Pkg; Pkg.add("Oceananigans")
 using OceanBioME, Oceananigans
 using Oceananigans.Units
 
-grid = RectilinearGrid(CPU(), size = (160, 32), extent = (500meters, 
-100meters), topology = (Bounded, Flat, Bounded))
+grid = RectilinearGrid(CPU(), size = (160, 32), extent = (10000meters, 500meters), topology = (Bounded, Flat, Bounded))
 
-biogeochemistry = NutrientPhytoplanktonZooplanktonDetritus(; grid, 
-open_bottom = true)
+biogeochemistry = NutrientPhytoplanktonZooplanktonDetritus(; grid) 
 
 model = NonhydrostaticModel(; grid, biogeochemistry,
                               advection = WENO(; grid),
-			                  buoyancy = SeawaterBuoyancy(constant_salinity = true),
-                              closure = AnisotropicMinimumDissipation())
+			                  closure = AnisotropicMinimumDissipation(),
+			                  buoyancy = SeawaterBuoyancy(constant_salinity = true))
 
-Tᵢ(x, y, z) = ifelse(x < grid.Lx/2, 8, 10)
+@inline front(x, z, μ, δ) = μ + δ * tanh((x - 7000 + 4 * z) / 500)
 
-set!(model, N = 5.0, P = 0.1, Z = 0.1, T = Tᵢ)
+Pᵢ(x, y, z) = ifelse(z > -50, 0.03, 0.01) 
+Nᵢ(x, y, z) = front(x, z, 2.5, -2)
+Tᵢ(x, y, z) = front(x, z, 9, 0.05)
 
-simulation = Simulation(model; Δt = 4.0, stop_time = 3hours)
+set!(model, N = Nᵢ, P = Pᵢ, Z = Pᵢ, T = Tᵢ)
 
-simulation.output_writers[:tracers] = JLD2OutputWriter(model, 
-model.tracers,
-                                                       filename = 
-"buoyancy_front.jld2",
-                                                       schedule = 
-TimeInterval(1minute),
-                                                       overwrite_existing 
-= true)
+simulation = Simulation(model; Δt = 50, stop_time = 4days)
+
+simulation.output_writers[:tracers] = JLD2OutputWriter(model, model.tracers,
+                                                       filename = "buoyancy_front.jld2",
+                                                       schedule = TimeInterval(24minute),
+                                                       overwrite_existing = true)
 
 run!(simulation)
 ```
@@ -66,10 +64,10 @@ We can then load the saved output and visualize it:
 
 ```@example quickstart
 T = FieldTimeSeries("buoyancy_front.jld2", "T")
+N = FieldTimeSeries("buoyancy_front.jld2", "N")
 P = FieldTimeSeries("buoyancy_front.jld2", "P")
 
-xT, yT, zT = nodes(T)
-xP, yP, zP = nodes(P)
+xc, yc, zc = nodes(T)
 
 times = T.times
 
@@ -77,28 +75,33 @@ using CairoMakie
 
 n = Observable(1)
 
-T_lims = (minimum(T), maximum(T))
-P_lims = (minimum(P), maximum(P))
+T_lims = (8.94, 9.06)
+N_lims = (0, 4.5)
+P_lims = (0.007, 0.02)
 
 Tₙ = @lift interior(T[$n], :, 1, :)
[…]
