### 2023-08-09 65e5652
added info line to work out why LOBSTER/InstantRemineralisation combination isn't running
