### 2023-08-10 70baf37
fixed set defaults for LOBSTER in sediment tests
