### 2023-08-10 OceanBioME/OceanBioME.jl#128 @iuryt
Add "Contributing" section to the Readme.md
Following one of the points in the checklist for JOSS
https://github.com/openjournals/joss-reviews/issues/5669#issuecomment-1667955949

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I suggest you to add a section to the Readme.md file similar to what Oceananigans.jl have.
https://github.com/CliMA/Oceananigans.jl#contributing
