### 2023-08-10 OceanBioME/OceanBioME.jl#129 @iuryt
State of the field
Following one of the points on the checklist
https://github.com/openjournals/joss-reviews/issues/5669#issuecomment-1667955949

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

I think that it should be added to the manuscript a description on how biogeochemical models are currently performed using Oceananigans.jl and how this package is improving the current state of the field.

Even myself have already used Oceananigans.jl coupled with a simple NP model, which led to a published paper,
https://www.frontiersin.org/articles/10.3389/fmars.2022.825027/full
https://github.com/iuryt/NorthAtlanticBloom
and I developed a very simple package that can be used for estimating the mixed-layer depth and light-limiting growth.
https://github.com/iuryt/Bioceananigans.jl

While those were a contribution to the field, I clearly see a huge improvement that your package is making to the state of the field, which I don't think you are making it clear in the manuscript.

## Advantages:

This is a long list, but in summary, this package is clearly well integrated with Oceananigans.jl and have many options for different models both using tracers and particles. It has a documentation and description of the different variables.

## Disadvantages (if any):

This is kind of hard to find a disadvantage for this package in relation to the current state of the field, but I will try to list something here.
The first one is that there is no way to change the equations or define my own set of equations for my biogeochemical model. This is important for theoretical process modelling approches.
I also could not find any way to define the mixed layer from the fields (there is a discussion to [add that to Oceanostics.jl](https://github.com/tomchor/Oceanostics.jl/issues/79)) and there are actually different ways to deal with the effect of the mixed layer on the biogeochemical modelling in the literature. The main point is that there are mismatches between the time scales of the mixing and cell division. Although light-averaging within the mixed layer has been previously employed and compared to observations (Mahadevan et al., 2012), other studies approach this issue differently (e.g., Sarmiento et al., 1993). Fasham et al., (1990), for example, average the light-limiting growth in the mixed layer. I could not see an implementation for that in this package.

While I don't think this is something you need to implement for the publication, it is important to, at least, mention that in the manuscript. If needed, I can add another issue for this improvement for tracking purposes.

## References

Mahadevan A., D’asaro E., Lee C., Perry M. J. (2012). Eddy-driven stratification initiates north atlantic spring phytoplankton blooms. Science 337, 54–58. doi: 10.1126/science.1218740

Sarmiento J.L., Slater R.D., Fasham M.J.R., Ducklow H.W., Toggweiler J.R., Evans G.T. (1993) A seasonal three‐dimensional ecosystem model of nitrogen cycling in the North Atlantic euphotic zone. Global biogeochemical cycles, 7 (2), 417–450. doi: 10.1029/93GB00375

Fasham M. J., Ducklow H. W., McKelvie S. M. (1990). A nitrogen-based model of plankton dynamics in the oceanic mixed layer. J. Mar. Res. 48, 591–639. doi: 10.1357/002224090784984678
