### 2023-08-11 8b2568c
Added contributors guide
