+++ README.md
@@ -134,3 +134,10 @@ grid = RectilinearGrid(GPU(), size = (256, 32), extent = (500meters, 100meters),
 
 See the [documentation](https://oceanbiome.github.io/OceanBioME.jl) for full description of the software
 package and more examples.
+
+## Contributing
+If you're interested in contributing to the development of OceanBioME we want would appreciate your help whatever contribution you make!
+
+If you'd like to work on a new feature, or if you're new to open source and want to crowd-source projects that fit your interests, you please start a discussion.
+
+For more information check out our [contributor's guide](https://oceanbiome.github.io/OceanBioME.jl/stable/contributing/).
