+++ docs/src/contributing.md
@@ -1,0 +1,166 @@
+# Contributors Guide
+
+Thank you for considering contributions to OceanBioME! We hope this guide
+helps.
+
+Feel free to ask us questions and chat with us at any time about any topic at all
+by:
+
+* [Opening a GitHub issue](https://github.com/OceanBioME/OceanBioME.jl/issues/new)
+
+* [Creating a GitHub discussion](https://github.com/OceanBioME/OceanBioME.jl/discussions/new)
+
+## Creating issues
+
+The simplest way to contribute to OceanBioME is to create or comment on issues and discussions.
+
+The most useful bug reports:
+
+* Provide an explicit code snippet --- not just a link --- that reproduces the bug in the latest tagged version of OceanBioME. This is sometimes called the ["minimal working example"](https://en.wikipedia.org/wiki/Minimal_working_example). Reducing bug-producing code to a minimal example can dramatically decrease the time it takes to resolve an issue.
+
+* Paste the _entire_ error received when running the code snippet, even if it's unbelievably long.
+
+* Use triple backticks (e.g., ````` ```some_code; and_some_more_code;``` `````) to enclose code snippets, and other [markdown formatting syntax](https://docs.github.com/en/github/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax) to make your issue easy and quick to read.
+
+* Report the OceanBioME version, Julia version, machine (especially if using a GPU) and any other possibly useful details of the computational environment in which the bug was created.
+
+Discussions are recommended for asking questions about (for example) the user interface, implementation details, science, and life in general.
+
+## But I want to _code_!
+
+* New users help write OceanBioME code and documentation by [forking the OceanBioME repository](https://docs.github.com/en/github/collaborating-with-pull-requests/working-with-forks), [using git](https://guides.github.com/introduction/git-handbook/) to edit code and docs, and then creating a [pull request](https://docs.github.com/en/github/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request-from-a-fork). Pull requests are reviewed by OceanBioME collaborators.
+
+* A pull request can be merged once it is reviewed and approved by collaborators. If the pull request author has write access, they have the responsibility of merging their pull request. Otherwise, OceanBioME.jl collaborators will execute the merge with permission from the pull request author.
+
+* Note: for small or minor changes (such as fixing a typo in documentation), the [GitHub editor](https://docs.github.com/en/github/managing-files-in-a-repository/managing-files-on-github/editing-files-in-your-repository) is super useful for forking and opening a pull request with a single click.
+
+* Write your code with love and care. In particular, conform to existing OceanBioME style and formatting conventions. For example, we love verbose and explicit variable names, use `TitleCase` for types, `snake_case` for objects, and always.put.spaces.after.commas. For formatting decisions we loosely follow the [YASGuide](https://github.com/jrevels/YASGuide). It's worth few extra minutes of our time to leave future generations with well-written, readable code.
+
+## What is a "collaborator" and how can I become one?
+
+* Collaborators have permissions to review pull requests and status allows a contributor to review pull requests in addition to opening them. Collaborators can also create branches in the main OceanBioME repository.
+
+* We ask that new contributors try their hand at forking OceanBioME, and opening and merging a pull request before requesting collaborator status.
+
+## What's a good way to start developing OceanBioME?
+
+* Tackle an existing issue. We keep a list of [good first issues](https://github.com/OceanBioME/OceanBioME.jl/issues?q=is%3Aissue+is%3Aopen+label%3A%22good+first+issue%22)
+  that are self-contained and suitable for a newcomer to try and work on.
+
[…]
