### 2023-08-15 7a2a647
Update README.md
fixed a few typos in README.md
