+++ README.md
@@ -136,8 +136,8 @@ See the [documentation](https://oceanbiome.github.io/OceanBioME.jl) for full des
 package and more examples.
 
 ## Contributing
-If you're interested in contributing to the development of OceanBioME we want would appreciate your help whatever contribution you make!
+If you're interested in contributing to the development of OceanBioME we would appreciate your help!
 
-If you'd like to work on a new feature, or if you're new to open source and want to crowd-source projects that fit your interests, you please start a discussion.
+If you'd like to work on a new feature, or if you're new to open source and want to crowd-source projects that fit your interests, please start a discussion.
 
 For more information check out our [contributor's guide](https://oceanbiome.github.io/OceanBioME.jl/stable/contributing/).
