+++ docs/src/model_components/biogeochemical/model_implementation.md
@@ -1,0 +1,333 @@
+# [Implementing a new models](@id model_implementation)
+
+Here we will describe how OceanBioME defines biogeochemical (BGC) models, how this varies from Oceananigans, and how to impliemnt your own model.
+
+## Model structure
+OceanBioME BGC models are `struct`s of type `ContinuousFormBiogeochemistry`, which is of abstract type `AbstractContinuousFormBiogeochemistry` from Oceananigans. In Oceananigans this describes BGC models which are continuous (depend continuously on ``x``, ``y``, and ``z``) rather than discrete (depending on ``i``, ``j``, ``k``). This simplifies the implementation of BGC models as forcing and sinking are very simple to define and add to tracers, and then Oceananigans handles the rest.
+
+OceanBioME's `ContinuousFormBiogeochemistry` adds a layer on top of this which makes adding [light attenuation models](@ref light), [sediment](@ref sediment), and [biologically active particles](@ref individuals). This is because `ContinuousFormBiogeochemistry` has parameters in which the types of these components are stored. This means that the model components will automatically be integrated with the BGC model without having to add new methods various Oceananigans functions as you otherwise will have. 
+
+## Implementing a model
+
+The nature of multiple dispatch in Julia means that we define new BGC models as new types. You can then define [methods](https://docs.julialang.org/en/v1/manual/methods/) to this type which are used by OceanBioME and Oceananigans to integrate the model.
+
+### The basics
+
+For this example we are going to implement the simple Nutrient-Phytoplankton model similar to that used in [](@citep), although we neglect the nutrient in/outflow terms since they may be added as [boundary conditions](https://clima.github.io/OceananigansDocumentation/stable/model_setup/boundary_conditions/), and modified to conserve nitrogen.
+
+The first step we need is to import the abstract type from OceanBioME, some units from Oceananigans (for ease of parameter definition), and [`import`](https://stackoverflow.com/questions/27086159/what-is-the-difference-between-using-and-import-in-julia-when-building-a-mod) some functions from Oceananigans which we will add methods to:
+
+```@example implementing
+using OceanBioME: ContinuousFormBiogeochemistry
+using Oceananigans.Units
+
+import Oceananigans.Biogeochemistry: required_biogeochemical_tracers,
+                                     required_biogeochemical_auxiliary_fields,
+                                     update_biogeochemical_state!,
+                                     biogeochemical_drift_velocity,
+				                     biogeochemical_auxiliary_fields
+```
+
+We then define our `struct` with the model parameters, as well as slots for the particles, light attenuation, and sediment models:
+
+```@example implementing
+@kwdef struct NutrientPhytoplankton{FT, LA, S, W, P} <:ContinuousFormBiogeochemistry{LA, S, P}
+            base_growth_rate :: FT = 1.27 / day              # 1 / seconds
+    nutrient_half_saturation :: FT = 0.025 * 1000 / 14       # mmol N / m³
+       light_half_saturation :: FT = 300.0                   # micro einstein / m² / s
+        temperature_exponent :: FT = 0.24                    # 1
[…]
