### 2023-08-16 5655a9b
Update NPZ.md (resize the image)
