+++ docs/oceanbiome.bib
@@ -263,7 +263,6 @@
 	year = {2015}
 }
 
-
 @article{RemineralisationFraction,
 	author = {Dunne, John P. and Sarmiento, Jorge L. and Gnanadesikan, Anand},
 	journal = {Global Biogeochemical Cycles},
@@ -272,3 +271,13 @@
 	volume = {21},
 	year = {2007}
 }
+
+
+@article{Chen2015,
+	author = {Ming Chen and Meng Fan and Rui Liu and Xiaoyu Wang and Xing Yuan and Huaiping Zhu},
+	journal = {Journal of Theoretical Biology},
+	month = {11},
+	pages = {8-19},
+	title = {The dynamics of temperature and light on the growth of phytoplankton},
+	volume = {385},
+	year = {2015}}
