+++ docs/src/model_components/biogeochemical/index.md
@@ -1,6 +1,6 @@
 # [Biogeochemical Models](@id bgc_models)
 
-Biogeochemical (BGC) models can be used within the [Oceananigans biogeochemistry framework](https://github.com/CliMA/Oceananigans.jl/pull/2802) or as stand alone box models. All BGC models should be setup in the same way so that they can easily be substituted for each other.
+Biogeochemical (BGC) models can be used within the [Oceananigans biogeochemistry framework](https://github.com/CliMA/Oceananigans.jl/pull/2802) or as stand alone box models. All BGC models should be setup in the same way so that they can easily be substituted for each other. You can easily implement a different model (or a variation on a current model) by following the guide [here](@ref bgc_model_implementation).
 
 For details of the BGC models currently implemented please see the following pages.
 
