### 2023-08-19 06ff6ea
Apply suggestions from code review
Co-authored-by: iury simoes-sousa <iuryt@pm.me>
