### 2023-08-19 12a927a
Update model_implementation.md
