### 2023-08-19 0eeccfb
changed image diaplay width limit
