### 2023-08-21 98d4883
Update NPZ.md
Co-authored-by: iury simoes-sousa <iuryt@pm.me>
