### 2023-08-23 caa3a52
Try to actually achieve the max width behaviour we want
