+++ docs/src/model_components/biogeochemical/NPZ.md
@@ -3,7 +3,7 @@
 The provided NPZD model is the low complexity model of [Kuhn2015](@citet).
 
 ```@raw html
-<img src="..npzd.svg"  alt="Diagram of NPZD formulation" style="max-width: calc(100% - 20px);">
+<img src="../npzd.svg"  alt="Diagram of NPZD formulation" style="max-width: calc(100% - 20px);">
 ```
 
 ## Model equations
