### 2023-08-24 d5a8022
added RK3 test back
