### 2023-08-25 ecad23a
updated docs example
