### 2023-08-25 e8137fb
Update model_implementation.md
Made some edits to the model implementation page to try to clarify a few aspects (and a few small corrections)
