### 2023-08-25 cab6e36
Apply suggestions from code review
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>
