### 2023-08-25 8e5e1ef
we still have to have the full depth sediment field to be able to output it
