### 2023-08-27 158936c
refactored models to abstract out light attenuation/sediments/particles more
