+++ src/Models/AdvectedPopulations/NPZD.jl
@@ -16,12 +16,12 @@ module NPZDModel
 
 export NutrientPhytoplanktonZooplanktonDetritus, NPZD
 
-using OceanBioME: ContinuousFormBiogeochemistry
+using OceanBioME: Biogeochemistry
 
 using Oceananigans.Units
 using Oceananigans.Fields: ZeroField
 
-using OceanBioME.Light: TwoBandPhotosyntheticallyActiveRadiation, update_PAR!, required_PAR_fields
+using OceanBioME.Light: TwoBandPhotosyntheticallyActiveRadiation
 using OceanBioME: setup_velocity_fields, show_sinking_velocities
 using OceanBioME.BoxModels: BoxModel
 using OceanBioME.Boundaries.Sediments: sinking_flux
@@ -31,7 +31,6 @@ import Base: show, summary
 
 import Oceananigans.Biogeochemistry: required_biogeochemical_tracers,
                                      required_biogeochemical_auxiliary_fields,
-                                     update_biogeochemical_state!,
                                      biogeochemical_drift_velocity,
 				                     biogeochemical_auxiliary_fields
 
@@ -41,7 +40,7 @@ import OceanBioME.Boundaries.Sediments: nitrogen_flux, carbon_flux, remineralisa
 
 import Adapt: adapt_structure, adapt
 
-struct NutrientPhytoplanktonZooplanktonDetritus{FT, LA, S, W, P} <: ContinuousFormBiogeochemistry{LA, S, P}
+struct NutrientPhytoplanktonZooplanktonDetritus{FT, W}
     # phytoplankton
     initial_photosynthetic_slope :: FT # α, 1/(W/m²)/s
     base_maximum_growth :: FT # μ₀, 1/s
@@ -59,16 +58,8 @@ struct NutrientPhytoplanktonZooplanktonDetritus{FT, LA, S, W, P} <: ContinuousFo
     # detritus
     remineralization_rate :: FT # rᵈⁿ, 1/s
 
-    # light attenuation
-    light_attenuation_model :: LA
-
-    # sediment
-    sediment_model :: S
-
     # sinking
     sinking_velocities :: W
-
-    particles :: P
 
     function NutrientPhytoplanktonZooplanktonDetritus(initial_photosynthetic_slope::FT,
                                                       base_maximum_growth::FT,
@@ -84,34 +75,22 @@ struct NutrientPhytoplanktonZooplanktonDetritus{FT, LA, S, W, P} <: ContinuousFo
     
                                                       remineralization_rate::FT,
     
-                                                      light_attenuation_model::LA,
-    
-                                                      sediment_model::S,
-    
-                                                      sinking_velocities::W,
-    
-                                                      particles::P) where {FT, LA, S, W, P}
-        return new{FT, LA, S, W, P}(initial_photosynthetic_slope,
-                                    base_maximum_growth,
-                                    nutrient_half_saturation,
-                                    base_respiration_rate,
-                                    phyto_base_mortality_rate,
+                                                      sinking_velocities::W) where {FT, W}
+        return new{FT, W}(initial_photosynthetic_slope,
+                          base_maximum_growth,
+                          nutrient_half_saturation,
+                          base_respiration_rate,
+                          phyto_base_mortality_rate,
 
-                                    maximum_grazing_rate,
-                                    grazing_half_saturation,
-                                    assimulation_efficiency,
-                                    base_excretion_rate,
-                                    zoo_base_mortality_rate,
+                          maximum_grazing_rate,
+                          grazing_half_saturation,
+                          assimulation_efficiency,
+                          base_excretion_rate,
+                          zoo_base_mortality_rate,
 
-                                    remineralization_rate,
-
-                                    light_attenuation_model,
-
-                                    sediment_model,
-
[…]
