+++ src/Models/AdvectedPopulations/LOBSTER/LOBSTER.jl
@@ -43,13 +43,11 @@ module LOBSTERModel
 
 export LOBSTER
 
-using OceanBioME: ContinuousFormBiogeochemistry
-
 using Oceananigans.Units
 using Oceananigans.Fields: Field, TracerFields, CenterField, ZeroField
 
-using OceanBioME.Light: TwoBandPhotosyntheticallyActiveRadiation, update_PAR!, required_PAR_fields
-using OceanBioME: setup_velocity_fields, show_sinking_velocities
+using OceanBioME.Light: TwoBandPhotosyntheticallyActiveRadiation
+using OceanBioME: setup_velocity_fields, show_sinking_velocities, Biogeochemistry
 using OceanBioME.BoxModels: BoxModel
 using OceanBioME.Boundaries.Sediments: sinking_flux
 
@@ -57,9 +55,7 @@ import OceanBioME.BoxModels: update_boxmodel_state!
 
 import Oceananigans.Biogeochemistry: required_biogeochemical_tracers,
                                      required_biogeochemical_auxiliary_fields,
-                                     biogeochemical_drift_velocity,
-                                     update_biogeochemical_state!,
-                                     biogeochemical_auxiliary_fields
+                                     biogeochemical_drift_velocity
 
 import OceanBioME: maximum_sinking_velocity
 
@@ -68,7 +64,7 @@ import Base: show, summary
 
 import OceanBioME.Boundaries.Sediments: nitrogen_flux, carbon_flux, remineralisation_receiver
 
-struct LOBSTER{FT, LA, S, B, W, P} <: ContinuousFormBiogeochemistry{LA, S, P}
+struct LOBSTER{FT, B, W}
     phytoplankton_preference :: FT
     maximum_grazing_rate :: FT
     grazing_half_saturation :: FT
@@ -99,14 +95,9 @@ struct LOBSTER{FT, LA, S, B, W, P} <: ContinuousFormBiogeochemistry{LA, S, P}
     disolved_organic_breakdown_rate :: FT
     zooplankton_calcite_dissolution :: FT
 
-    light_attenuation_model :: LA
-    sediment_model :: S
-
     optionals :: B
 
     sinking_velocities :: W
-
-    particles :: P
 
     function LOBSTER(phytoplankton_preference::FT,
                      maximum_grazing_rate::FT,
@@ -138,53 +129,43 @@ struct LOBSTER{FT, LA, S, B, W, P} <: ContinuousFormBiogeochemistry{LA, S, P}
                      disolved_organic_breakdown_rate::FT,
                      zooplankton_calcite_dissolution::FT,
 
-                     light_attenuation_model::LA,
-                     sediment_model::S,
-
                      optionals::B,
                 
-                     sinking_velocities::W,
-                     
-                     particles::P) where {FT, LA, S, B, W, P}
+                     sinking_velocities::W) where {FT, B, W}
 
-        return new{FT, LA, S, B, W, P}(phytoplankton_preference,
-                                       maximum_grazing_rate,
-                                       grazing_half_saturation,
-                                       light_half_saturation,
-                                       nitrate_ammonia_inhibition,
-                                       nitrate_half_saturation,
-                                       ammonia_half_saturation,
-                                       maximum_phytoplankton_growthrate,
-                                       zooplankton_assimilation_fraction,
-                                       zooplankton_mortality,
-                                       zooplankton_excretion_rate,
-                                       phytoplankton_mortality,
-                                       small_detritus_remineralisation_rate,
-                                       large_detritus_remineralisation_rate,
[…]
