### 2023-08-27 cb6cdfa
updated light attenuation test
