### 2023-08-27 9a8fc8e
updated light attenuation test
