+++ src/Models/AdvectedPopulations/LOBSTER/LOBSTER.jl
@@ -412,11 +412,8 @@ adapt_structure(to, lobster::LOBSTER) =
             lobster.fast_sinking_mortality_fraction,
             lobster.disolved_organic_breakdown_rate,
             lobster.zooplankton_calcite_dissolution,
-            adapt(to, lobster.light_attenuation_model),
-            adapt(to, lobster.sediment_model),
             lobster.optionals,
-            adapt(to, lobster.sinking_velocities), 
-            adapt(to, lobster.particles))
+            adapt(to, lobster.sinking_velocities))
 
 summary(::LOBSTER{FT, B, W}) where {FT, B, W} = string("Lodyc-DAMTP Ocean Biogeochemical Simulation Tools for Ecosystem and Resources (LOBSTER) model ($FT)")
 show(io::IO, model::LOBSTER{FT, Val{B}, W}) where {FT, B, W} =
