### 2023-08-27 5d7d019
added redfield function
