### 2023-08-27 cb01e55
updated examples to use redfield function
