### 2023-08-27 a15377a
added some docstrings and moved `underlying_model` to arg (rather than kwarg
