### 2023-08-27 da33077
fixed stuff and added show/summary
