### 2023-08-27 0b2def6
fixed doctests (hopefully again)
