### 2023-08-27 b88d638
Fixed other LOBSTER test function (oops)
