### 2023-08-27 def4adf
fixed examples (again)
