+++ Project.toml
@@ -12,12 +12,12 @@ Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
 StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
 
 [compat]
-Adapt = "3.4"
-JLD2 = "0.4.31"
+Adapt = "3"
+JLD2 = "^0.4"
 KernelAbstractions = "0.9"
-Oceananigans = "0.84.1, 0.85"
+Oceananigans = "0.84.1, 0.85, 0.86, 0.87"
 Roots = "2"
-StructArrays = "0.6.15"
+StructArrays = "0.4, 0.5, 0.6"
 julia = "1.9"
 
 [extras]
