### 2023-08-28 53ff3b3
Update eady.jl
reduced initial timestep to 1 minute in Eady.jl to be more conservative and avoid CFL violation on finer grids
