### 2023-08-29 6f89d73
Update examples/eady.jl
