### 2023-08-29 1f87d4d
Update eady.jl
