### 2023-08-29 1721923
fixed show stuff
