+++ src/OceanBioME.jl
@@ -148,7 +148,7 @@ Returns the redfield ratio of `tracer_name` from `bgc` when it is constant acros
 
 summary(bgc::Biogeochemistry) = string("Biogeochemical model based on $(summary(bgc.underlying_biogeochemistry))")
 show(io::IO, model::Biogeochemistry) =
-       print(io, show(io, model.underlying_biogeochemistry), " \n",
+       print(io, show(model.underlying_biogeochemistry), " \n",
                 " Light attenuation: ", summary(model.light_attenuation), "\n",
                 " Sediment: ", summary(model.sediment), "\n",
                 " Particles: ", summary(model.particles))
