+++ src/Boundaries/Sediments/Sediments.jl
@@ -3,26 +3,20 @@ module Sediments
 export SimpleMultiG, InstantRemineralisation
 
 using KernelAbstractions
-<<<<<<< HEAD
-=======
-using OceanBioME: ContinuousFormBiogeochemistry, BoxModelGrid
->>>>>>> origin
+
+using OceanBioME: Biogeochemistry, BoxModelGrid
+
 using Oceananigans
 using Oceananigans.Architectures: device
 using Oceananigans.Utils: launch!
 using Oceananigans.Advection: advective_tracer_flux_z
 using Oceananigans.Units: day
-using Oceananigans.Fields: CenterField, Face
+using Oceananigans.Fields: CenterField, Face, Center, ConstantField
 using Oceananigans.Biogeochemistry: biogeochemical_drift_velocity
 using Oceananigans.Grids: zspacing
 using Oceananigans.Operators: volume
-<<<<<<< HEAD
-using Oceananigans.Fields: Center
-using OceanBioME: Biogeochemistry
-=======
-using Oceananigans.Fields: Center, ConstantField
 using Oceananigans.ImmersedBoundaries: ImmersedBoundaryGrid, immersed_cell
->>>>>>> origin
+
 
 import Adapt: adapt_structure, adapt
 import Oceananigans.Biogeochemistry: update_tendencies!
