### 2023-08-30 ea49268
Update model_implementation.md
