### 2023-08-30 eb90278
finished 'modifier' framework
