### 2023-08-30 7e0b694
updated zero negative tracers to be more reliable
