### 2023-08-30 1f6a4c1
added method to automatically setup negative tracer scalking
