+++ src/OceanBioME.jl
@@ -150,6 +150,13 @@ Returns the redfield ratio of `tracer_name` from `bgc` when it is constant acros
 @inline redfield(i, j, k, val_tracer_name, bgc::Biogeochemistry, tracers) = redfield(i, j, k, val_tracer_name, bgc.underlying_biogeochemistry, tracers)
 @inline redfield(val_tracer_name, bgc::Biogeochemistry) = redfield(val_tracer_name, bgc.underlying_biogeochemistry)
 
+"""
+    conserved_tracers(model::UnderlyingBiogeochemicalModel)
+
+Returns the names of tracers which together are conserved in `model`
+"""
+conserved_tracers() = nothing
+
 summary(bgc::Biogeochemistry) = string("Biogeochemical model based on $(summary(bgc.underlying_biogeochemistry))")
 show(io::IO, model::Biogeochemistry) =
        print(io, show(model.underlying_biogeochemistry), " \n",
