### 2023-08-30 ca4cf2e
updated negative protection tests
