+++ examples/data_forced.jl
@@ -106,13 +106,11 @@ simulation.output_writers[:profiles] = JLD2OutputWriter(model,
 
 # TODO: make tendency callback to force no NaNs in tendencies
 
-scale_negative_tracers = ScaleNegativeTracers(; model, tracers = (:NO₃, :NH₄, :P, :Z, :sPOM, :bPOM, :DOM))
-simulation.callbacks[:neg] = Callback(scale_negative_tracers; callsite = UpdateStateCallsite())
-
 wizard = TimeStepWizard(cfl = 0.2, diffusive_cfl = 0.2,
                         max_change = 1.5, min_change = 0.75,
                         cell_diffusion_timescale = column_diffusion_timescale,
                         cell_advection_timescale = column_advection_timescale)
+                        
 simulation.callbacks[:wizard] = Callback(wizard, IterationInterval(10))
 nothing #hide
 
