### 2023-08-31 3a71aaf
hide output after loading exampel resoltes
