### 2023-08-31 0deb0a1
simplified LOBSTER tests
