+++ test/test_LOBSTER.jl
@@ -1,18 +1,11 @@
 using Test
-using OceanBioME: LOBSTER
+using OceanBioME: LOBSTER, conserved_tracers, redfield
 using Oceananigans, CUDA
 
-ΣN(model, variable_redfield) = variable_redfield ? (
-    sum(model.tracers.NO₃) + sum(model.tracers.NH₄) + sum(model.tracers.P) + sum(model.tracers.Z) + sum(model.tracers.sPON) + sum(model.tracers.bPON) + sum(model.tracers.DON)) : (
-    sum(model.tracers.NO₃) + sum(model.tracers.NH₄) + sum(model.tracers.P) + sum(model.tracers.Z) + sum(model.tracers.sPOM) + sum(model.tracers.bPOM) + sum(model.tracers.DOM))
+ΣN(model, biogeochemistry) = sum([sum(model.tracers[tracer_name]) for tracer_name in conserved_tracers(biogeochemistry)])
 
-function ΣC(model, carbonates, variable_redfield)
-    # *will only be conserved if carbonates on*
-    if variable_redfield
-        OC = sum(model.tracers.sPOC .+ model.tracers.bPOC .+ model.tracers.DOC)
-    else
-        OC = sum(model.tracers.sPOM .+ model.tracers.bPOM .+ model.tracers.DOM) * model.biogeochemistry.underlying_biogeochemistry.organic_redfield
-    end
+function ΣC(model, carbonates, biogeochemistry)
+    OC = sum([sum(model.tracers[tracer_name] * redfield(Val(tracer_name), biogeochemistry, model.tracers)) for tracer_name in conserved_tracers(biogeochemistry)])
 
     if carbonates
         IC = sum(model.tracers.DIC)
@@ -20,23 +13,13 @@ function ΣC(model, carbonates, variable_redfield)
         IC = 0.0
     end
 
-    LC = sum(model.tracers.P * (1 + model.biogeochemistry.underlying_biogeochemistry.organic_carbon_calcate_ratio) .+ model.tracers.Z) * model.biogeochemistry.underlying_biogeochemistry.phytoplankton_redfield 
-
-    return OC + IC + LC
+    return OC + IC
 end
 
-ΣGⁿ(model, variable_redfield) = variable_redfield ? (
-    sum(model.timestepper.Gⁿ.NO₃) + sum(model.timestepper.Gⁿ.NH₄) + sum(model.timestepper.Gⁿ.P) + sum(model.timestepper.Gⁿ.Z) + sum(model.timestepper.Gⁿ.sPON) + sum(model.timestepper.Gⁿ.bPON) + sum(model.timestepper.Gⁿ.DON)) : (
-    sum(model.timestepper.Gⁿ.NO₃) + sum(model.timestepper.Gⁿ.NH₄) + sum(model.timestepper.Gⁿ.P) + sum(model.timestepper.Gⁿ.Z) + sum(model.timestepper.Gⁿ.sPOM) + sum(model.timestepper.Gⁿ.bPOM) + sum(model.timestepper.Gⁿ.DOM))
+ΣGⁿ(model, biogeochemistry) = sum([sum(model.timestepper.Gⁿ[tracer_name]) for tracer_name in conserved_tracers(biogeochemistry)])
 
-
-function ΣGᶜ(model, carbonates, variable_redfield)
-    # *will only be conserved if carbonates on*
-    if variable_redfield
-        OC = sum(model.timestepper.Gⁿ.sPOC .+ model.timestepper.Gⁿ.bPOC .+ model.timestepper.Gⁿ.DOC)
-    else
[…]
