### 2023-08-31 ba0d49f
Updated bgc docs page
