+++ docs/src/model_components/biogeochemical/index.md
@@ -5,7 +5,7 @@ Biogeochemical (BGC) models can be used within the [Oceananigans biogeochemistry
 For details of the BGC models currently implemented please see the following pages.
 
 ## Oceananigans setup
-At the simplest level all that is required to setup a BGC model is to pass it to the Oceananigans model setup:
+At the simplest level all that is required to setup an existint OceanBioME BGC model is to pass it to the Oceananigans model setup:
 ```julia
 model = NonhydrostaticModel(; grid,
                               ...,
@@ -19,3 +19,7 @@ MODEL_NAME(; grid, growth_rate = 10.0)
 This will set up the required tracers and auxiliary fields, and you may also set boundary conditions or additional forcing through the normal Oceananigans setup. 
 
 Models usually have a default [light attenuation model](@ref light) specified, these may be substituted easily by passing different models as parameters as above.
+
+Our models are implemented in an abstract framework `Biogeochemistry` which contains `underlying_biogeochemistry`, `light_attenuation`, `sediment`, and `modifiers`.
+This is automatically setup for existing BGC models, but may also be used to couple any BGC model with light attenuation and sediments.
+See the [implementation page](@ref model_implementation) for some more information on how to couple other models.
