### 2023-09-02 0191dc1
dropped `UnderlyingBiogeochemistry` in favor of better using the Oceananigans Asbtract types
This means any Ocenanigans BGC can go in the slot, and we don't need to define some extra things
