### 2023-09-02 a9dac4b
updated model implementation page
