+++ paper/paper.bib
@@ -161,3 +161,18 @@
    year = 2022,
    doi = {10.17226/26278}
 }
+
+@article{taylor:2016,
+	author = {Taylor, J. R.},
+   doi = {10.1002/2016GL069106},
+	issn = {19448007},
+	issue = {11},
+	journal = {Geophysical Research Letters},
+   publisher = {Blackwell Publishing Ltd},
+	month = {6},
+   pages = {5784-5792},
+   volume = {43},
+   bdsk-url-1 = {https://doi.org/10.1002/2016GL069106},
+	title = {Turbulent mixing, restratification, and phytoplankton growth at a submesoscale eddy},
+	year = {2016}
+}
