### 2023-09-06 #5669 @jagoosw
> @jagoosw The new implementations to the package and documentation are awesome! Let me know when you have an updated version of the manuscript and I can finish the review.

Thank you! The current version of the PDF (above) contains the changes you suggested.
