### 2023-09-07 af43c8c
Apply suggestions from code review
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>
