+++ src/OceanBioME.jl
@@ -139,6 +139,7 @@ Returns the redfield ratio of `tracer_name` from `bgc` at `i`, `j`, `k`.
 @inline redfield(i, j, k, val_tracer_name, bgc, tracers) = redfield(val_tracer_name, bgc, tracers)
 
 """
+    redfield(val_tracer_name, bgc)
     redfield(val_tracer_name, bgc, tracers)
 
 Returns the redfield ratio of `tracer_name` from `bgc` when it is constant across the domain.
