+++ examples/column.jl
@@ -42,9 +42,9 @@ grid = RectilinearGrid(size=(1, 1, 50), extent=(20meters, 20meters, 200meters))
 
 
 # ## Model
-# First we will define the biogeochemical model including carbonate chemistry 
+# First we define the biogeochemical model including carbonate chemistry 
 # and scaling of negative tracers(see discussion in the [positivity preservation](@ref pos-preservation))
-# and then setup the Oceananigans model with the boundary condition for the DIC based on the air-sea CO₂ flux
+# and then setup the Oceananigans model with the boundary condition for the DIC based on the air-sea CO₂ flux.
 
 biogeochemistry = LOBSTER(; grid,
                             surface_phytosynthetically_active_radiation = PAR⁰,
