+++ src/Light/Light.jl
@@ -26,7 +26,6 @@ import Base: show, summary
 import Oceananigans.Biogeochemistry: biogeochemical_auxiliary_fields, update_biogeochemical_state!, required_biogeochemical_auxiliary_fields
 import Oceananigans.BoundaryConditions: _fill_top_halo!
 
-# Fallback
 include("2band.jl")
 include("morel.jl")
 
