### 2023-09-07 ad12a21
moved `Pkg.add`s out of `@example` blocks in quick start to see if that fixes the docs build issue
