+++ docs/src/quick_start.md
@@ -3,7 +3,7 @@
 OceanBioME provides biogeochemical models to plug into [Oceananigans](https://github.com/CliMA/Oceananigans.jl), for example this code will run one month of a single column, 7 variable (P, Z, sPOM, bPOM, DOM, NO₃, NH₄) biogeochemical situation with constant forcing.
 
 First we need to check we have the required dependencies:
-```@example quickstart
+```julia
 using Pkg
 Pkg.add(["OceanBioME", "Oceananigans"])
 ```
