### 2023-09-07 6583027
fixed gpu incompatability in `calculate_bottom_indices`
