### 2023-09-07 20524a9
fixed adapt method for instant remineralisation (oops)
