+++ .github/workflows/tests.yml
@@ -17,9 +17,8 @@ env:
 
 jobs:
   test:
+    name: Julia ${{ matrix.version }} - self-hosted - ${{ matrix.arch }} - ${{ github.event_name }}
     runs-on: self-hosted
-    name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
-    runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
       matrix:
