### 2023-09-08 ed8bc4c
tried to make a slurm script
