+++ test/gpu_runtests.jl
@@ -1,0 +1,13 @@
+using Pkg
+
+Pkg.instantiate()
+
+Pkg.add("Test", "CUDA", "DataDeps", "Documenter", "Statistics", "JLD2")
+
+Pkg.precompile()
+
+using Oceananigans
+
+arch = GPU()
+
+include("runtests.jl")
