+++ test/runtests.jl
@@ -1,5 +1,9 @@
 using OceanBioME, Documenter, Test
 
+if !(@isdefined arch)
+    arch = CPU()
+end
+
 include("test_utils.jl")
 include("test_light.jl")
 include("test_LOBSTER.jl")
@@ -8,6 +12,8 @@ include("test_gasexchange.jl")
 include("test_slatissima.jl")
 include("test_sediments.jl")
 
-@testset "Doctests" begin
-    doctest(OceanBioME)
+if isa(arch, CPU)
+    @testset "Doctests" begin
+        doctest(OceanBioME)
+    end
 end

[…]
