### 2023-09-08 42cbb3e
maybe fixed CPU test
