### 2023-09-08 e066bf7
Reverted cpu tests to github runner
