+++ .github/workflows/gpu-tests.yml
@@ -23,13 +23,10 @@ jobs:
     name: Deploy CSD3 tests
     runs-on: self-hosted
     steps:
-      - label: "initialize gpu enviroment"
-      commands:
-        - "wget -N -P $CSD3_HOME/_temp https://julialang-s3.julialang.org/bin/linux/x64/$JULIA_MINOR_VERSION/julia-$JULIA_VERSION-linux-x86_64.tar.gz"
-        - "tar xf $CSD3_HOME/_temp/julia-$JULIA_VERSION-linux-x86_64.tar.gz -C $CSD3_HOME/_temp"
-
-        # Run the tests on slurm
-        - "sbatch .github/workflows/gpu-test.sbatch"
-
-        # Clean up
-        - "rm -rf $CSD3_HOME/_temp/julia-$JULIA_VERSION" # presumably this happens anyway
+      - name: Download julia"
+        run: "wget -N -P $CSD3_HOME/_temp https://julialang-s3.julialang.org/bin/linux/x64/$JULIA_MINOR_VERSION/julia-$JULIA_VERSION-linux-x86_64.tar.gz"
+        run: "tar xf $CSD3_HOME/_temp/julia-$JULIA_VERSION-linux-x86_64.tar.gz -C $CSD3_HOME/_temp"
+      - name: Run tests
+        run: "sbatch .github/workflows/gpu-test.sbatch"
+      - name: Clean up
+        run: "rm -rf $CSD3_HOME/_temp/julia-$JULIA_VERSION" # presumably this happens anyway
