+++ .github/workflows/gpu-tests.yml
@@ -23,10 +23,11 @@ jobs:
     name: Deploy CSD3 tests
     runs-on: self-hosted
     steps:
-      - name: Download julia"
+      - name: Download Julia
         run: "wget -N -P $RDS_HOME/_temp https://julialang-s3.julialang.org/bin/linux/x64/$JULIA_MINOR_VERSION/julia-$JULIA_VERSION-linux-x86_64.tar.gz;
               tar xf $RDS_HOME/_temp/julia-$JULIA_VERSION-linux-x86_64.tar.gz -C $RDS_HOME/_temp"
       - name: Run tests
-        run: "sbatch .github/workflows/gpu-test.sbatch"
+        run: "sbatch .github/workflows/gpu-test.sbatch;
+              /home/js2430/rds/hpc-work/OceanBioME-runner/_work/_temp/julia-1.9.3/bin/julia -O0 --color=yes --project .github/workflows/check_results.jl"
       - name: Clean up
-        run: "rm -rf $RDS_HOME/_temp/julia-$JULIA_VERSION" # presumably this happens anyway
+        run: "rm -rf $RDS_HOME/_temp/julia-$JULIA_VERSION; rm results.test" # presumably this happens anyway
