### 2023-09-08 1c95287
Delete .github/workflows/check_results.jl
