+++ .github/workflows/gpu-tests.yml
@@ -32,4 +32,5 @@ jobs:
         run: "sbatch .github/workflows/gpu-test.sbatch;
               /home/js2430/rds/hpc-work/OceanBioME-runner/_work/_temp/julia-1.9.3/bin/julia -O0 --color=yes --project test/check_results.jl"
       - name: Clean up
+        if: always()
         run: "rm -rf $RDS_HOME/_temp/julia-$JULIA_VERSION; rm results.test" # presumably this happens anyway
