### 2023-09-08 140835f
Remove output too
