### 2023-09-09 OceanBioME/OceanBioME.jl#129 @iuryt
Hi @jagoosw
This issue is closed, but I just noticed that it was not checked in the task list from #126
I also couldn't find a "State of the field" section or mentioned in the text of the updated version of the manuscript.
