### 2023-09-11 125cf0a
added GPU detection to tests
