### 2023-09-12 #5669 @kthyng
@iuryt I believe the next step in this process is for you to check out the newest manuscript PDF (apologies if my glancing through the review issue is incorrect).
