### 2023-09-12 OceanBioME/OceanBioME.jl#129 @jagoosw
Hi @iuryt,

We resolved most of this issue in #131 but I had forgotten to add to the manuscript. I have done this now!

Thanks
