### 2023-09-12 #5669 @jagoosw
I believe that all of the links go to the current version, you can see each version in the action results in our PR https://github.com/OceanBioME/OceanBioME.jl/pull/76, e.g. you can download `paper` at the bottom of this https://github.com/OceanBioME/OceanBioME.jl/actions/runs/6100131308 from the prior version
