### 2023-09-13 OceanBioME/OceanBioME.jl#140 @iuryt
Figure 1 from JOSS paper
This should be an easy fix.

As mentioned in https://github.com/openjournals/joss-reviews/issues/5669#issuecomment-1666968906

The axis tick labels in this figure is unnecessarily small. If you cannot find a way to make it larger, you can remove it and mention the dimensions of the domain in the caption.

All other fonts from the images are also small, but not as small as these tick labels.
I usually try to use fonts around the same size as the main text.

![image](https://github.com/OceanBioME/OceanBioME.jl/assets/5797727/d144cf4e-4cee-4d53-a4e5-be278ed48d7a)
