### 2023-09-14 OceanBioME/OceanBioME.jl#140 @jagoosw
Were working on running this in higher resolution so I will update when that's done!
