### 2023-09-14 038bfaf
Update paper.md
Various edits to the text
