### 2023-09-14 1464028
Update paper.md
Added mention of Oceananigans w/o OceanBioME and reference to Simoes-Sousa
