### 2023-09-14 6343f00
Update paper/paper.md
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>
