### 2023-09-14 f591649
Update paper/paper.md
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>
