### 2023-09-14 34c1574
Update paper/paper.md
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>
