+++ src/Boundaries/Sediments/Sediments.jl
@@ -37,7 +37,7 @@ function update_tendencies!(bgc, sediment::FlatSediment, model)
 
     launch!(arch, model.grid, :xy,
             _calculate_tendencies!,
-            bgc.sediment, bgc, model.grid, model.advection, model.tracers, model.timestepper.Gⁿ)
+            sediment, bgc.underlying_biogeochemistry, model.grid, model.advection, model.tracers, model.timestepper)
             
     return nothing
 end
