+++ src/Models/AdvectedPopulations/NPZD.jl
@@ -199,7 +199,7 @@ function NutrientPhytoplanktonZooplanktonDetritus(; grid,
                                                  sinking_velocities)
 
     if scale_negatives
-        scaler = ScaleNegativeTracers(underlying_biogeochemistry)
+        scaler = ScaleNegativeTracers(underlying_biogeochemistry, grid)
         modifiers = isnothing(modifiers) ? scaler : (modifiers..., scaler)
     end
 
