+++ paper/paper.bib
@@ -176,3 +176,12 @@
 	title = {Turbulent mixing, restratification, and phytoplankton growth at a submesoscale eddy},
 	year = {2016}
 }
+@article{@Simoes-Sousa2022,
+  title={Mixed layer eddies supply nutrients to enhance the spring phytoplankton bloom},
+  author={Simoes-Sousa, Iury T and Tandon, Amit and Pereira, Filipe and Lazaneo, Caue Z and Mahadevan, Amala},
+  journal={Frontiers in Marine Science},
+  volume={9},
+  pages={825027},
+  year={2022},
+  publisher={Frontiers}
+}
