+++ paper/paper.bib
@@ -183,5 +183,6 @@
   volume={9},
   pages={825027},
   year={2022},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi = {10.3389/fmars.2022.825027}
 }
