### 2023-09-14 c5c9569
fixed utils test
