+++ paper/paper.bib
@@ -176,7 +176,7 @@
 	title = {Turbulent mixing, restratification, and phytoplankton growth at a submesoscale eddy},
 	year = {2016}
 }
-@article{Simoes-Sousa2022,
+@article{SimoesSousa2022,
   title={Mixed layer eddies supply nutrients to enhance the spring phytoplankton bloom},
   author={Simoes-Sousa, Iury T and Tandon, Amit and Pereira, Filipe and Lazaneo, Caue Z and Mahadevan, Amala},
   journal={Frontiers in Marine Science},
