### 2023-09-15 a0d8da4
fixed npzd and losbter tests
