### 2023-09-15 058619b
hunch about memory not being freed
