### 2023-09-15 1ff75e5
forgot to remove a show (but also revealed that a test wasn't using GPU
