+++ test/test_gasexchange.jl
@@ -52,7 +52,7 @@ end
 end
 
 @testset "Gas exchange coupling" begin
-    grid = RectilinearGrid(size=(1, 1, 2), extent=(1, 1, 1))
+    grid = RectilinearGrid(architecture; size=(1, 1, 2), extent=(1, 1, 1))
     conc_field = CenterField(grid, indices=(:, :, grid.Nz))
     conc_field .= 413.0 + 1.0 * rand()
     conc_function(x, y, t) = 413.0 + 10.0 * sin(t * π / (1year))
