### 2023-09-15 40f0d28
removed scalar indexing from kelp tests
