### 2023-09-15 64fc80e
removed scalar indexing from sediment tests
