+++ test/test_gasexchange.jl
@@ -24,8 +24,10 @@ function test_gas_exchange_model(grid, air_concentration)
 
     set!(model, T = 15.0, S = 35.0, DIC = 2220, Alk = 2500)
 
-    # is everything communicating properly?
-    @test Oceananigans.getbc(model.tracers.DIC.boundary_conditions.top, 1, 1, grid, model.clock, fields(model)) ≈ -0.0003 atol = 0.0001
+    # is everything communicating properly? (can't think of a way to not use allow scalar here)
+    value = CUDA.@allowscalar Oceananigans.getbc(model.tracers.DIC.boundary_conditions.top, 1, 1, grid, model.clock, fields(model))
+
+    @test value ≈ -0.0003 atol = 0.0001
 
     # just incase we broke something
     @test isnothing(time_step!(model, 1.0))
