### 2023-09-15 49282b3
don't run doctests on GPU
