+++ test/test_gasexchange.jl
@@ -53,11 +53,12 @@ end
     @test (mean(pCO₂_err) < 10 && std(pCO₂_err) < 15)
 end
 
+@inline conc_function(x, y, t) = 413.0 + 10.0 * sin(t * π / (1year))
+
 @testset "Gas exchange coupling" begin
     grid = RectilinearGrid(architecture; size=(1, 1, 2), extent=(1, 1, 1))
     conc_field = CenterField(grid, indices=(:, :, grid.Nz))
     conc_field .= 413.0 + 1.0 * rand()
-    conc_function(x, y, t) = 413.0 + 10.0 * sin(t * π / (1year))
     for air_concentration in [413.1, conc_function, conc_field]
         @info "Testing with $(typeof(air_concentration))"
         test_gas_exchange_model(grid, air_concentration)
