### 2023-09-16 b32322e
Update paper.md
fixed typo for "n prep" that somehow didn't get changed in merged pull request
