### 2023-09-16 #5669 @iuryt
@kthyng
It was a great pleasure to review this paper! I learned a lot and did my best to give as many contributions as I could as a reviewer.
The package has a great potential in the field for being very easy to run and nicely integrated with the physical model.
Even I want to use it in some of the current projects I am working and I know a couple of other scientists that are also interested in using it for their research, which means that this is a fresh package with a high demand.
The documentation also explains how to develop new modules, which means that users can implement their own equations if they want. This is particularly important for theoretical process modelling initiatives.
The fact that it is easy to run and the possibility of running it as a box model also shows a great potential for use in class while teaching biogeochemical interactions.
They are also showing that their package presents a huge improvement from the current-used methods with the same model.
Based on all of these comments and with the new improvements in the documentation and the paper, I recommend the publication of the paper/package as is.

My only final suggestion for the author's proof stage, and this is mentioned in OceanBioME/OceanBioME.jl#140 , is that the figures could have larger fontsize for the labels for accessibility.
