### 2023-09-16 e414c19
fixed gas exchange on fields
