+++ src/Boundaries/gasexchange.jl
@@ -3,6 +3,8 @@
 ##### As per OCMIP Phase 2 http://ocmip5.ipsl.jussieu.fr/OCMIP/phase2/simulations/Abiotic/Cchem/co2calc.f simplified as in PISCESv2
 #####
 
+using Oceananigans.Fields: interpolate
+
 struct pCO₂{P0, P1, P2, PB, PW, FT}
                   solubility :: P0
     bicarbonate_dissociation :: P1
@@ -257,9 +259,4 @@ end
 
 @inline get_value(x, y, t, air_concentration::Number) = air_concentration
 @inline get_value(x, y, t, air_concentration::Function) = air_concentration(x, y, t)
-
-# interpolation does not work on 2d grids as trilinear interpolation fails but we're only ever going to be asking for values on grid points
-@inline function get_value(x, y, t, air_concentration::Field{Center, Center, Center})
-    (ξ, i), (η, j), (ζ, k) = modf.(fractional_indices(x, y, znodes(air_concentration)[end], (Center(), Center(), Center()), air_concentration.grid))
-    return air_concentration[floor(Int, i) + 1, floor(Int, j) + 1, floor(Int, k) + 1]
-end
+@inline get_value(x, y, t, conc::Field{Center, Center, Center}) = interpolate(conc, Center(), Center(), Center(), conc.grid, x, y, 0)
