### 2023-09-17 670076a
gas exhcnage boundary value fix
