+++ test/test_gasexchange.jl
@@ -62,7 +62,7 @@ grid = RectilinearGrid(architecture; size=(1, 1, 2), extent=(1, 1, 1))
 #Oceananigans.fill_halo_regions!(conc_field)
 
 @testset "Gas exchange coupling" begin
-    for air_concentration in [413.1, conc_function, conc_field]
+    for air_concentration in [413.1, conc_function]#, conc_field]
         @info "Testing with $(typeof(air_concentration))"
         test_gas_exchange_model(grid, air_concentration)
     end
