### 2023-09-18 9f93ebe
corrected kelp test tollerance for GPU
