+++ src/Boundaries/Sediments/instant_remineralization.jl
@@ -100,7 +100,7 @@ sediment_fields(model::InstantRemineralisation) = (N_storage = model.fields.N_st
     # sediment evolution
     @inbounds sediment.tendencies.Gⁿ.N_storage[i, j, 1] = burial_efficiency * flux
 
-    @inbounds tendencies[remineralisation_receiver(bgc)][i, j, 1] += flux * (1 - burial_efficiency) / Δz
+    @inbounds tendencies[remineralisation_receiver(bgc)][i, j, k] += flux * (1 - burial_efficiency) / Δz
 end
 
 summary(::InstantRemineralisation{FT}) where {FT} = string("Single-layer instant remineralisaiton ($FT)")
