### 2023-09-18 ac93375
added eady script and changed resolution in text
