+++ paper/figures/eady.jl
@@ -1,0 +1,130 @@
+using OceanBioME, Printf, Oceananigans, Oceananigans.Units
+
+using Oceananigans.Architectures: arch_array
+
+using Oceananigans.Solvers: FFTBasedPoissonSolver 
+
+using Random
+
+Random.seed!(11)
+
+h(k) = (k - 1) / Nz
+
+ζ₀(k) = 1 + (h(k) - 1) / refinement
+
+Σ(k) = (1 - exp(-stretching * h(k))) / (1 - exp(-stretching))
+
+z_faces(k) = Lz * (ζ₀(k) * Σ(k) - 1)
+
+Nz = 64
+Lz = 140  
+
+refinement = 1.8
+stretching = 3
+
+arch = GPU()
+
+grid = RectilinearGrid(arch; size = (512, 512, Nz), x = (0, 1kilometer), y = (0, 1kilometer), z = z_faces)
+
+Lx, Ly, Lz = grid.Lz, grid.Ly, grid.Lz
+
+coriolis = FPlane(f = 1e-4) # [s⁻¹]
+
+background_state_parameters = ( M = 1e-4,       # s⁻¹, geostrophic shear
+                                f = coriolis.f, # s⁻¹, Coriolis parameter
+                                N = 1e-4,       # s⁻¹, buoyancy frequency
+                                H = grid.Lz )
+
+# We assume a background buoyancy ``B`` with a constant stratification and also a constant lateral
+# gradient (in the zonal direction). The background velocity components ``U`` and ``V`` are prescribed
+# so that the thermal wind relationship is satisfied, that is, ``f \partial_z U = - \partial_y B`` and
+# ``f \partial_z V = \partial_x B``.
+B(x, y, z, t, p) = p.M ^ 2 * x + p.N ^ 2 * (z + p.H)
+V(x, y, z, t, p) = p.M ^ 2 / p.f * (z + p.H)
+
+V_field = BackgroundField(V, parameters = background_state_parameters)
+B_field = BackgroundField(B, parameters = background_state_parameters)
+
+νᵥ = κᵥ = 1e-4 # [m² s⁻¹]
+vertical_diffusivity = VerticalScalarDiffusivity(ν = νᵥ, κ = κᵥ)
+
+@info "Setting up kelp particles"
+
+n = 32
+
+x = arch_array(arch, [repeat([-200.0, -100.0, 0.0, 100.0, 200.0], 1, 5)...] .+ Lx / 2)
+y = arch_array(arch, [repeat([-200.0, -100.0, 0.0, 100.0, 200.0], 1, 5)'...] .+ Ly / 2)
+z = arch_array(arch, zeros(Float64, n))
+
+particles = SLatissima(; architecture = arch,
+                         x, y, z, 
+                         A = arch_array(arch, 5.0 .* ones(n)), N = arch_array(arch, 0.01.* ones(n)), C = arch_array(arch, 0.18.* ones(n)), 
+                         latitude = 57.5,
+                         scalefactor = 10.0 ^ 2,
+                         pescribed_temperature = (args...) -> 12.0)
+
+biogeochemistry = LOBSTER(; grid,
+                            carbonates = true,
+                            open_bottom = true,
+                            particles,
+			    scale_negatives = true)
+
+DIC_bcs = FieldBoundaryConditions(top = GasExchange(; gas = :CO₂, temperature = (args...) -> 12, salinity = (args...) -> 35))
+
+# Model instantiation
+model = NonhydrostaticModel(; grid,
+                              biogeochemistry,
+                              boundary_conditions = (DIC = DIC_bcs, ),
+                              advection = WENO(grid),
+                              timestepper = :RungeKutta3,
+                              coriolis,
+                              tracers = :b,
+                              buoyancy = BuoyancyTracer(),
[…]
