### 2023-09-18 54404a9
Documenter 1.0 just dropped and it breaks a lot of stuff that I will fix in a different PR
