+++ Project.toml
@@ -19,6 +19,7 @@ Oceananigans = "0.84.1, 0.85, 0.86, 0.87"
 Roots = "2"
 StructArrays = "0.4, 0.5, 0.6"
 julia = "1.9"
+Documenter = "0.27"
 
 [extras]
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
