### 2023-09-18 30cdfba
tried to simplify multi G adapted to GPU [skip ci]
