+++ src/Utils/negative_tracers.jl
@@ -111,7 +111,7 @@ end
         end
     end 
 
-    t < 0 && error("Cell total < 0, can not scale negative tracers.")
+    t < 0 && t = NaN
 
     @unroll for tracer in tracers
         value = @inbounds tracer[i, j, k]
