### 2023-09-18 19a40f1
removed column diffusion stuff
