### 2023-09-18 fb35422
missed an allow scalar
