+++ docs/Project.toml
@@ -18,3 +18,4 @@ CairoMakie = "0.10"
 DocumenterCitations = "1"
 Literate = "≥2.9.0"
 Oceananigans = "0.84.1, 0.85, 0.86, 0.87"
+Documenter = "0.27"
