### 2023-09-18 61e10f4
missed an allow scalar
