### 2023-09-18 31cb1b7
use simpler LOBSTER model with Multi G to fit parameter space
