### 2023-09-18 ebea1e0
lat/lon test still not working properly
