### 2023-09-19 98e47e6
remove column diffusion timescale references
