### 2023-09-19 b133ef0
corrected run criteria on GPU
