### 2023-09-19 37a21c1
corrected run criteria on GPU
