### 2023-09-19 ba39a35
hopefully reduced sediment kernel parameter size
