+++ src/Boundaries/Sediments/simple_multi_G.jl
@@ -55,11 +55,10 @@ end
                    fast_fraction = 0.74,
                    slow_fraction = 0.26,
                    refactory_fraction = 0.1,
-                   nitrate_oxidation_params = (A = - 1.9785, B = 0.2261, C = -0.0615, D = -0.0289, E = - 0.36109, F = - 0.0232),
-                   denitrification_params = (A = - 3.0790, B = 1.7509, C = 0.0593, D = - 0.1923, E = 0.0604, F = 0.0662),
-                   anoxic_params = (A = - 3.9476, B = 2.6269, C = - 0.2426, D = -1.3349, E = 0.1826, F = - 0.0143),
-                   depth = abs(znodes(grid, Face())[1]),
-                   solid_dep_params = (A = 0.233, B = 0.336, C = 982, D = - 1.548, depth = depth))
+                   nitrate_oxidation_params = (- 1.9785, 0.2261, -0.0615, -0.0289, - 0.36109, - 0.0232),
+                   denitrification_params = (- 3.0790, 1.7509, 0.0593, - 0.1923, 0.0604, 0.0662),
+                   anoxic_params = (- 3.9476, 2.6269, - 0.2426, -1.3349, 0.1826, - 0.0143),
+                   solid_dep_params = (0.233, 0.336, 982.0, - 1.548))
 
 Return a single-layer "multi G" sediment model (`SimpleMultiG`) on `grid`, where parameters
 can be optionally specified.
@@ -97,30 +96,10 @@ function SimpleMultiG(; grid,
                        fast_fraction = 0.74,
                        slow_fraction = 0.26,
                        refactory_fraction = 0.1,
-                       nitrate_oxidation_params = (A = - 1.9785,
-                                                       B = 0.2261,
-                                                       C = -0.0615,
-                                                       D = -0.0289,
-                                                       E = - 0.36109,
-                                                       F = - 0.0232),
-                       denitrification_params = (A = - 3.0790,
-                                                     B = 1.7509,
-                                                     C = 0.0593,
-                                                     D = - 0.1923,
-                                                     E = 0.0604,
-                                                     F = 0.0662),
-                       anoxic_params = (A  = - 3.9476,
-                                            B = 2.6269,
-                                            C = - 0.2426,
-                                            D = -1.3349,
-                                            E = 0.1826,
-                                            F = - 0.0143),
-                       depth = abs(znodes(grid, Face())[1]),
-                       solid_dep_params = (A = 0.233, 
-                                               B = 0.336, 
-                                               C = 982, 
-                                               D = - 1.548,
-                                               depth = depth))
+                       nitrate_oxidation_params = (- 1.9785, 0.2261, -0.0615, -0.0289, - 0.36109, - 0.0232),
+                       denitrification_params = (- 3.0790, 1.7509, 0.0593, - 0.1923, 0.0604, 0.0662),
+                       anoxic_params = (- 3.9476, 2.6269, - 0.2426, -1.3349, 0.1826, - 0.0143),
+                       solid_dep_params = (0.233, 0.336, 982.0, - 1.548))
 
[…]
