### 2023-09-20 b9b42b3
improved generality of `adapt_structure` methods
