+++ src/Models/Individuals/SLatissima.jl
@@ -201,54 +201,54 @@ Base.@kwdef struct SLatissima{AR, FT, U, T, S, P, F} <: BiogeochemicalParticles
     latitude :: FT = 57.5
 end
 
-adapt_structure(to, kelp::SLatissima) = SLatissima(kelp.architecture,
-                                                   kelp.growth_rate_adjustement, 
-                                                   kelp.photosynthetic_efficiency,
-                                                   kelp.minimum_carbon_reserve,
-                                                   kelp.structural_carbon,
-                                                   kelp.exudation,
-                                                   kelp.erosion,
-                                                   kelp.saturation_irradiance,
-                                                   kelp.structural_dry_weight_per_area,
-                                                   kelp.structural_dry_to_wet_weight,
-                                                   kelp.carbon_reserve_per_carbon,
-                                                   kelp.nitrogen_reserve_per_nitrogen,
-                                                   kelp.minimum_nitrogen_reserve,
-                                                   kelp.maximum_nitrogen_reserve,
-                                                   kelp.growth_adjustement_2,
-                                                   kelp.growth_adjustement_1,
-                                                   kelp.maximum_specific_growth_rate,
-                                                   kelp.structural_nitrogen,
-                                                   kelp.photosynthesis_at_ref_temp_1,
-                                                   kelp.photosynthesis_at_ref_temp_2,
-                                                   kelp.photosynthesis_ref_temp_1,
-                                                   kelp.photosynthesis_ref_temp_2,
-                                                   kelp.photoperiod_1,
-                                                   kelp.photoperiod_2,
-                                                   kelp.respiration_at_ref_temp_1,
-                                                   kelp.respiration_at_ref_temp_2,
-                                                   kelp.respiration_ref_temp_1,
-                                                   kelp.respiration_ref_temp_2,
-                                                   kelp.photosynthesis_arrhenius_temp,
-                                                   kelp.photosynthesis_low_temp,
-                                                   kelp.photosynthesis_high_temp,
-                                                   kelp.photosynthesis_high_arrhenius_temp,
-                                                   kelp.photosynthesis_low_arrhenius_temp,
-                                                   kelp.respiration_arrhenius_temp,
-                                                   kelp.current_speed_for_0p65_uptake,
-                                                   kelp.nitrate_half_saturation,
-                                                   kelp.ammonia_half_saturation,
-                                                   kelp.maximum_nitrate_uptake,
-                                                   kelp.maximum_ammonia_uptake,
-                                                   kelp.current_1,
-                                                   kelp.current_2,
-                                                   kelp.current_3,
-                                                   kelp.respiration_reference_A,
-                                                   kelp.respiration_reference_B,
-                                                   kelp.exudation_redfield_ratio,
-                                                   kelp.pescribed_velocity,
-                                                   kelp.pescribed_temperature,
-                                                   kelp.pescribed_salinity,
+adapt_structure(to, kelp::SLatissima) = SLatissima(adapt(to, kelp.architecture),
+                                                   adapt(to, kelp.growth_rate_adjustement), 
+                                                   adapt(to, kelp.photosynthetic_efficiency),
+                                                   adapt(to, kelp.minimum_carbon_reserve),
+                                                   adapt(to, kelp.structural_carbon),
+                                                   adapt(to, kelp.exudation),
+                                                   adapt(to, kelp.erosion),
+                                                   adapt(to, kelp.saturation_irradiance),
+                                                   adapt(to, kelp.structural_dry_weight_per_area),
+                                                   adapt(to, kelp.structural_dry_to_wet_weight),
[…]
