+++ src/Models/AdvectedPopulations/LOBSTER/LOBSTER.jl
@@ -400,36 +400,36 @@ function update_boxmodel_state!(model::BoxModel{<:Biogeochemistry{<:LOBSTER}, <:
 end
 
 adapt_structure(to, lobster::LOBSTER) =
-    LOBSTER(lobster.phytoplankton_preference,
-            lobster.maximum_grazing_rate,
-            lobster.grazing_half_saturation,
-            lobster.light_half_saturation,
-            lobster.nitrate_ammonia_inhibition,
-            lobster.nitrate_half_saturation,
-            lobster.ammonia_half_saturation,
-            lobster.maximum_phytoplankton_growthrate,
-            lobster.zooplankton_assimilation_fraction,
-            lobster.zooplankton_mortality,
-            lobster.zooplankton_excretion_rate,
-            lobster.phytoplankton_mortality,
-            lobster.small_detritus_remineralisation_rate,
-            lobster.large_detritus_remineralisation_rate,
-            lobster.phytoplankton_exudation_fraction,
-            lobster.nitrifcaiton_rate,
-            lobster.ammonia_fraction_of_exudate,
-            lobster.ammonia_fraction_of_excriment,
-            lobster.ammonia_fraction_of_detritus,
-            lobster.phytoplankton_redfield,
-            lobster.organic_redfield,
-            lobster.phytoplankton_chlorophyll_ratio,
-            lobster.organic_carbon_calcate_ratio,
-            lobster.respiraiton_oxygen_nitrogen_ratio,
-            lobster.nitrifcation_oxygen_nitrogen_ratio,
-            lobster.slow_sinking_mortality_fraction,
-            lobster.fast_sinking_mortality_fraction,
-            lobster.disolved_organic_breakdown_rate,
-            lobster.zooplankton_calcite_dissolution,
-            lobster.optionals,
+    LOBSTER(adapt(to, lobster.phytoplankton_preference),
+            adapt(to, lobster.maximum_grazing_rate),
+            adapt(to, lobster.grazing_half_saturation),
+            adapt(to, lobster.light_half_saturation),
+            adapt(to, lobster.nitrate_ammonia_inhibition),
+            adapt(to, lobster.nitrate_half_saturation),
+            adapt(to, lobster.ammonia_half_saturation),
+            adapt(to, lobster.maximum_phytoplankton_growthrate),
+            adapt(to, lobster.zooplankton_assimilation_fraction),
+            adapt(to, lobster.zooplankton_mortality),
+            adapt(to, lobster.zooplankton_excretion_rate),
+            adapt(to, lobster.phytoplankton_mortality),
+            adapt(to, lobster.small_detritus_remineralisation_rate),
+            adapt(to, lobster.large_detritus_remineralisation_rate),
+            adapt(to, lobster.phytoplankton_exudation_fraction),
+            adapt(to, lobster.nitrifcaiton_rate),
+            adapt(to, lobster.ammonia_fraction_of_exudate),
+            adapt(to, lobster.ammonia_fraction_of_excriment),
+            adapt(to, lobster.ammonia_fraction_of_detritus),
+            adapt(to, lobster.phytoplankton_redfield),
+            adapt(to, lobster.organic_redfield),
+            adapt(to, lobster.phytoplankton_chlorophyll_ratio),
+            adapt(to, lobster.organic_carbon_calcate_ratio),
+            adapt(to, lobster.respiraiton_oxygen_nitrogen_ratio),
+            adapt(to, lobster.nitrifcation_oxygen_nitrogen_ratio),
+            adapt(to, lobster.slow_sinking_mortality_fraction),
+            adapt(to, lobster.fast_sinking_mortality_fraction),
[…]
