### 2023-09-20 2cdee7b
removed unneccesary `Array(interior(...))`
