### 2023-09-21 2dfb211
updated model implementation page
