### 2023-09-23 6b7b3fb
Remove redundant fallbacks
