### 2023-09-24 a566ab1
updated eady figure
