+++ paper/figures/eady.jl
@@ -6,7 +6,7 @@ using Oceananigans.Solvers: FFTBasedPoissonSolver
 
 using Random
 
-Random.seed!(11)
+Random.seed!(43)
 
 h(k) = (k - 1) / Nz
 
@@ -16,17 +16,15 @@ h(k) = (k - 1) / Nz
 
 z_faces(k) = Lz * (ζ₀(k) * Σ(k) - 1)
 
-Nz = 64
-Lz = 140  
+Nx, Ny, Nz = 512, 512, 64
+Lx, Ly, Lz = 1kilometer, 1kilometer, 140.0
 
 refinement = 1.8
 stretching = 3
 
 arch = GPU()
 
-grid = RectilinearGrid(arch; size = (512, 512, Nz), x = (0, 1kilometer), y = (0, 1kilometer), z = z_faces)
-
-Lx, Ly, Lz = grid.Lz, grid.Ly, grid.Lz
+grid = RectilinearGrid(arch; size = (Nx, Ny, Nz), x = (0, Lx), y = (0, Ly), z = z_faces)
 
 coriolis = FPlane(f = 1e-4) # [s⁻¹]
 
@@ -46,28 +44,29 @@ V_field = BackgroundField(V, parameters = background_state_parameters)
 B_field = BackgroundField(B, parameters = background_state_parameters)
 
 νᵥ = κᵥ = 1e-4 # [m² s⁻¹]
-vertical_diffusivity = VerticalScalarDiffusivity(ν = νᵥ, κ = κᵥ)
+closure = ScalarDiffusivity(ν = νᵥ, κ = κᵥ)
 
 @info "Setting up kelp particles"
 
-n = 32
+n = 36 # must be a square number
 
-x = arch_array(arch, [repeat([-200.0, -100.0, 0.0, 100.0, 200.0], 1, 5)...] .+ Lx / 2)
-y = arch_array(arch, [repeat([-200.0, -100.0, 0.0, 100.0, 200.0], 1, 5)'...] .+ Ly / 2)
+x = arch_array(arch, [repeat([Lx / (sqrt(n) + 1) * n for n in 1:Int(sqrt(n))], 1, Int(sqrt(n)))...])
+y = arch_array(arch, [repeat([Ly / (sqrt(n) + 1) * n for n in 1:Int(sqrt(n))], 1, Int(sqrt(n)))'...])
 z = arch_array(arch, zeros(Float64, n))
 
 particles = SLatissima(; architecture = arch,
                          x, y, z, 
-                         A = arch_array(arch, 5.0 .* ones(n)), N = arch_array(arch, 0.01.* ones(n)), C = arch_array(arch, 0.18.* ones(n)), 
-                         latitude = 57.5,
-                         scalefactor = 10.0 ^ 2,
+                         A = arch_array(arch, 5.0 .* ones(n)), N = arch_array(arch, 0.01 .* ones(n)), C = arch_array(arch, 0.18 .* ones(n)), 
+                         latitude = 43.3,
+                         scalefactor = 500.0,
                          pescribed_temperature = (args...) -> 12.0)
 
 biogeochemistry = LOBSTER(; grid,
                             carbonates = true,
                             open_bottom = true,
                             particles,
-			    scale_negatives = true)
+                            scale_negatives = true,
+			    surface_phytosynthetically_active_radiation = (x, y, t) -> 100)
 
 DIC_bcs = FieldBoundaryConditions(top = GasExchange(; gas = :CO₂, temperature = (args...) -> 12, salinity = (args...) -> 35))
 
@@ -81,7 +80,7 @@ model = NonhydrostaticModel(; grid,
                               tracers = :b,
                               buoyancy = BuoyancyTracer(),
                               background_fields = (b = B_field, v = V_field),
[…]

[…]
