### 2023-09-24 8a70f46
added sentence about color ordering
