### 2023-09-24 526efab
added citation for KA and CUDA
