+++ paper/paper.bib
@@ -186,3 +186,21 @@
   publisher={Frontiers},
   doi = {10.3389/fmars.2022.825027}
 }
+@ARTICLE{CUDA,
+   author={Besard, Tim and Foket, Christophe and De Sutter, Bjorn},
+   journal={IEEE Transactions on Parallel and Distributed Systems}, 
+   title={Effective Extensible Programming: Unleashing Julia on GPUs}, 
+   year={2019},
+   volume={30},
+   number={4},
+   pages={827-841},
+   doi={10.1109/TPDS.2018.2872064}
+}
+@article{KA,
+   title={JuliaGPU/KernelAbstractions.jl: v0.9.8}, 
+   DOI={10.5281/zenodo.8176369}, 
+   publisher={Zenodo}, 
+   author={Valentin Churavy and Dilum Aluthge and Anton Smirnov and Julian Samaroo and James Schloss and Lucas C Wilcox and Simon Byrne and Tim Besard and Maciej Waruszewski and Ali Ramadhan and et al.}, 
+   year={2023}, 
+   month={Jul}
+}
