### 2023-09-24 69eb822
Apply suggestions from code review
Co-authored-by: johnryantaylor <johnryantaylor@gmail.com>
