### 2023-09-24 c3506d2
Update src/Boundaries/Sediments/simple_multi_G.jl
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>
