### 2023-09-24 675d256
included Julia GPU in acknowledgements
