### 2023-09-24 79a58aa
Apply suggestions from code review
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>
