### 2023-09-24 210951b
Apply suggestion from code review
Co-authored-by: Navid C. Constantinou <navidcy@users.noreply.github.com>

This got removed by the last PR
