+++ src/Boundaries/Sediments/simple_multi_G.jl
@@ -29,17 +29,18 @@ end
 
 """
     SimpleMultiG(; grid
-                   fast_decay_rate = 2/day,
-                   slow_decay_rate = 0.2/day,
-                   fast_redfield = 0.1509,
-                   slow_redfield = 0.13,
-                   fast_fraction = 0.74,
-                   slow_fraction = 0.26,
-                   refactory_fraction = 0.1,
-                   nitrate_oxidation_params = (- 1.9785, 0.2261, -0.0615, -0.0289, - 0.36109, - 0.0232),
-                   denitrification_params = (- 3.0790, 1.7509, 0.0593, - 0.1923, 0.0604, 0.0662),
-                   anoxic_params = (- 3.9476, 2.6269, - 0.2426, -1.3349, 0.1826, - 0.0143),
-                   solid_dep_params = (0.233, 0.336, 982.0, - 1.548))
+                   fast_decay_rate::FT = 2/day,
+                   slow_decay_rate::FT = 0.2/day,
+                   fast_redfield::FT = 0.1509,
+                   slow_redfield::FT = 0.13,
+                   fast_fraction::FT = 0.74,
+                   slow_fraction::FT = 0.26,
+                   refactory_fraction::FT = 0.1,
+                   nitrate_oxidation_params::P1 = (A = - 1.9785, B = 0.2261, C = -0.0615, D = -0.0289, E = - 0.36109, F = - 0.0232),
+                   denitrification_params::P2 = (A = - 3.0790, B = 1.7509, C = 0.0593, D = - 0.1923, E = 0.0604, F = 0.0662),
+                   anoxic_params::P3 = (A = - 3.9476, B = 2.6269, C = - 0.2426, D = -1.3349, E = 0.1826, F = - 0.0143),
+                   depth = abs(znodes(grid, Face())[1]),
+                   solid_dep_params::P4 = (A = 0.233, B = 0.336, C = 982, D = - 1.548, depth = depth))
 
 Return a single-layer "multi G" sediment model (`SimpleMultiG`) on `grid`, where parameters
 can be optionally specified.
@@ -70,17 +71,37 @@ Single-layer multi-G sediment model (Float64)
 ```
 """
 function SimpleMultiG(; grid,
-                       fast_decay_rate = 2/day,
-                       slow_decay_rate = 0.2/day,
-                       fast_redfield = 0.1509,
-                       slow_redfield = 0.13,
-                       fast_fraction = 0.74,
-                       slow_fraction = 0.26,
-                       refactory_fraction = 0.1,
-                       nitrate_oxidation_params = (- 1.9785, 0.2261, -0.0615, -0.0289, - 0.36109, - 0.0232),
-                       denitrification_params = (- 3.0790, 1.7509, 0.0593, - 0.1923, 0.0604, 0.0662),
-                       anoxic_params = (- 3.9476, 2.6269, - 0.2426, -1.3349, 0.1826, - 0.0143),
-                       solid_dep_params = (0.233, 0.336, 982.0, - 1.548))
+                       fast_decay_rate::FT = 2/day,
+                       slow_decay_rate::FT = 0.2/day,
+                       fast_redfield::FT = 0.1509,
+                       slow_redfield::FT = 0.13,
+                       fast_fraction::FT = 0.74,
+                       slow_fraction::FT = 0.26,
+                       refactory_fraction::FT = 0.1,
+                       nitrate_oxidation_params::P1 = (A = - 1.9785,
+                                                       B = 0.2261,
+                                                       C = -0.0615,
+                                                       D = -0.0289,
[…]
