### 2023-09-25 745501e
bump oceananigans compat entry
