+++ Project.toml
@@ -15,7 +15,7 @@ StructArrays = "09ab397b-f2b6-538f-b94a-2f83cf4a842a"
 Adapt = "3"
 JLD2 = "0.4"
 KernelAbstractions = "0.9"
-Oceananigans = "0.84.1, 0.85, 0.86, 0.87"
+Oceananigans = "0.84.1, 0.85, 0.86, 0.87, 0.88"
 Roots = "2"
 StructArrays = "0.4, 0.5, 0.6"
 julia = "1.9"
