### 2023-09-25 a8bed1e
updated eady figure Co-authored-by: johnryantaylor <johnryantaylor@gmail.com>
