### 2023-09-26 OceanBioME/OceanBioME.jl#140 @glwagner
Found it:

![](https://raw.githubusercontent.com/OceanBioME/OceanBioME.jl/a8bed1ede046fd580438c02c4432ec37435fd26a/paper/eady_example.png)
