### 2023-09-27 e4258c1
Update paper.md
small edits to the description in the caption to Fig 1
