+++ paper/paper.md
@@ -75,7 +75,7 @@ On top of this, we added particles modelling the growth of sugar kelp, which are
 Thanks to Julia's speed and efficiency the above model (1 km × 1 km × 140 m with 512 × 512 × 64 grid points) took about 2 hours of computing time to simulate 10 days of evolution on an Nvidia A100 GPU. 
 The figure shows the domain with the colour representing the concentration of various biogeochemical tracer fields: nutrients (nitrate and ammonia), phytoplankton, organic matter (dissolved and particulate), and dissolved inorganic carbon. 
 Darker colours represent higher values.
-High concentrations of organic matter develop near the center of the eddy. Small-scale variability and enhanced subduction can also be seen along a sub-mesoscale front.
+High concentrations of organic matter develop near the centre of the eddy. Small-scale variability and enhanced subduction can also be seen along a sub-mesoscale front.
 Points on the surface represent the kelp particle positions, with the colour representing the range of frond size.
 Figure made with ``Makie.jl`` [@makie]. \label{eady}](eady_example.png)
 
@@ -153,6 +153,6 @@ $^6$ `SLatissima`
 
 # Acknowledgements
 
-We would like to thank the [Climate Modeling Alliance](https://clima.caltech.edu) team and ``Oceananigans.jl`` contributors for their fantastic project. We also thank ``JuliaGPU`` contributors for ``KernelAbstractions.jl`` and ``CUDA.jl`` which make this code possible. Additionally, we are very grateful for the support and funding of the [Centre for Climate Repair, Cambridge](https://www.climaterepair.cam.ac.uk/) and the [Gordon and Betty Moore Foundation](https://www.moore.org/). N.C.C. acknowledges support by the Australian Research Council DECRA Fellowship DE210100749.
+We want to thank the [Climate Modeling Alliance](https://clima.caltech.edu) team and ``Oceananigans.jl`` contributors for their fantastic project. We also thank ``JuliaGPU`` contributors for ``KernelAbstractions.jl`` and ``CUDA.jl`` which make this code possible. Additionally, we are very grateful for the support and funding of the [Centre for Climate Repair, Cambridge](https://www.climaterepair.cam.ac.uk/) and the [Gordon and Betty Moore Foundation](https://www.moore.org/). N.C.C. acknowledges support from the Australian Research Council DECRA Fellowship DE210100749. Finally, we would like to thank the reviewers Iury Simoes-Sousa and Francis Poulin for their very helpful comments. 
 
 # References
