### 2023-10-02 #5669 @kthyng
Hi @jagoosw ,and yep! Ping me here when you have released 0.7.0 and I'll include it here. I'll go ahead and put in the list of things I'll need you to do next, also.
