+++ src/TimeSteppers/TimeSteppers.jl
@@ -1,7 +1,0 @@
-module TimeSteppers
-
-using Oceananigans.TimeSteppers: AbstractTimeStepper
-
-include("positive_conservative_AB2.jl")
-
-end
