+++ src/TimeSteppers/positive_conservative_AB2.jl
@@ -1,162 +1,0 @@
-using Oceananigans.Fields: FunctionField, location
-using Oceananigans.TurbulenceClosures: implicit_step!
-using Oceananigans.Utils: @apply_regionally, apply_regionally!
-
-mutable struct PositiveQuasiAdamsBashforth2TimeStepper{FT, GT, IT} <: AbstractTimeStepper
-                  χ :: FT
-        previous_Δt :: FT
-                 Gⁿ :: GT
-                 G⁻ :: GT
-    implicit_solver :: IT
-  intermediate_step :: GT
-end
-
-"""
-PositiveQuasiAdamsBashforth2TimeStepper(grid, tracers,
-                                    χ = 0.1;
-                                    implicit_solver = nothing,
-                                    Gⁿ = TendencyFields(grid, tracers),
-                                    G⁻ = TendencyFields(grid, tracers))
-
-Return a 2nd-order quasi Adams-Bashforth (AB2) time stepper (`QuasiAdamsBashforth2TimeStepper`)
-on `grid`, with `tracers`, and AB2 parameter `χ`. The tendency fields `Gⁿ` and `G⁻` can be
-specified via  optional `kwargs`.
-
-The 2nd-order quasi Adams-Bashforth timestepper steps forward the state `Uⁿ` by `Δt` via
-
-```julia
-Uⁿ⁺¹ = Uⁿ + Δt * [(3/2 + χ) * Gⁿ - (1/2 + χ) * Gⁿ⁻¹]
-```
-
-where `Uⁿ` is the state at the ``n``-th timestep, `Gⁿ` is the tendency
-at the ``n``-th timestep, and `Gⁿ⁻¹` is the tendency at the previous
-timestep (`G⁻`).
-
-!!! note "First timestep"
-    For the first timestep, since there are no saved tendencies from the previous timestep,
-    the `QuasiAdamsBashforth2TimeStepper` performs an Euler timestep:
-
-    ```julia
-    Uⁿ⁺¹ = Uⁿ + Δt * Gⁿ
-    ```
-"""
-function PositiveQuasiAdamsBashforth2TimeStepper(grid, tracers,
-                                         χ = 0.1;
-                                         implicit_solver::IT = nothing,
-                                         Gⁿ = TendencyFields(grid, tracers),
-                                         G⁻ = TendencyFields(grid, tracers),
-                                         Cⁿ⁺¹ = TendencyFields(grid, tracers)) where IT
-
-    FT = eltype(grid)
-    GT = typeof(Gⁿ)
-
-    return PositiveQuasiAdamsBashforth2TimeStepper{FT, GT, IT}(χ, Inf, Gⁿ, G⁻, implicit_solver, Cⁿ⁺¹)
-end
-
-function reset!(timestepper::PositiveQuasiAdamsBashforth2TimeStepper)
-    timestepper.previous_Δt = Inf
-    return nothing
-end
-
-#####
-##### Time steppping
-#####
-
-"""
-    time_step!(model::AbstractModel{<:PositiveQuasiAdamsBashforth2TimeStepper}, Δt; euler=false)
-
-Step forward `model` one time step `Δt` with a 2nd-order Adams-Bashforth method and
-pressure-correction substep. Setting `euler=true` will take a forward Euler time step.
-"""
-function time_step!(model::AbstractModel{<:PositiveQuasiAdamsBashforth2TimeStepper}, Δt; callbacks = [], euler=false)
-    Δt == 0 && @warn "Δt == 0 may cause model blowup!"
-
-    # Shenanigans for properly starting the AB2 loop with an Euler step
-    euler = euler || (Δt != model.timestepper.previous_Δt)
-    
-    χ = ifelse(euler, convert(eltype(model.grid), -0.5), model.timestepper.χ)
-
-    if euler
[…]
