### 2023-10-04 #5669 @kthyng
@jagoosw can you update the metadata for your zenodo archive to match your JOSS paper title and author list? This is not a strict requirement but it looks good and clarifies that the JOSS paper is specifically for that version.
