### 2023-10-05 #5669 @jagoosw
Done now (I think)!

Just to let you know when I click the link in this:

> Done! archive is now [v0.7.0](https://doi.org/v0.7.0)

It goes to the wrong place.
