### 2023-10-05 59a581e
remove duplicate Documenter compat entry
