+++ docs/Project.toml
@@ -19,4 +19,3 @@ Documenter = "0.27"
 DocumenterCitations = "1"
 Literate = "≥2.9.0"
 Oceananigans = "0.84.1, 0.85, 0.86, 0.87"
-Documenter = "0.27"
