### 2023-10-10 #5669 @johnryantaylor
Hi @kthyng, thank you and @iuryt and @francispoulin very much for your help this paper.  Is there a chance that I could add a line to the acknowledgements section?  I realized today that I failed to acknowledge the Kelp Forest Foundation which did some of the legwork in helping us get funded for this.  It would be wondeful if I could add a sentence like the following to the acknowledgements section: "The authors would like to thank the Kelp Forest Foundation for their help in getting this work started and for many helpful discussions."  I understand if its too late, but I thought I would ask!
