### 2023-10-10 #5669 @kthyng
Hi @johnryantaylor! Yes go ahead and modify the paper file and regenerate to make sure it looks how you want. Then I can re-accept the paper.
