### 2023-10-10 #5669 @navidcy
> Hi @johnryantaylor! Yes go ahead and modify the paper file and regenerate to make sure it looks how you want. Then I can re-accept the paper.

@kthyng, the original branch `joss-paper` is now merged. Should we make another branch with the same name and modify the paper there?
