### 2023-10-10 #5669 @kthyng
No need to make a new branch, you can just edit the paper in the main branch if you want, and no need for a new version since it's all just paper edits. I can update the branch on my end.
