### 2023-10-13 #5669 @kthyng
Ok that command doesn't work after the issue is closed, but I can go ahead and reaccept. Let me know if it doesn't look right and I can reaccept again.
